==> 02_datasetup_QC.Rmd <==
---
title: "QC and Setup of NEO data"
author: "David Hughes"
date: "20/8/2021"
output: 
  pdf_document:
    number_sections: true
---

This PDF report will define the merging and setup of the data as well as the quality control steps of the covariate data.

The metabolite data has previously been QC'd by metaboprep


```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.width = 4, fig.height = 4, echo = TRUE, error = FALSE, warning = FALSE, message = FALSE)
```

## Read in the data

```{r}
### parameter file
pfile = "parameters/pfile.txt"
par_data = read.table(pfile, header = FALSE, sep = "=", as.is = TRUE)

### Read in the source data
w = grep("pheno_file", par_data[,1])
f = paste0(par_data[w,2] )
sourcedata = read.table(f, header = TRUE, sep = ",", as.is = TRUE, fill = FALSE)
sourcedata$ID = paste0("ind", sourcedata$ID)

## Read in the GRS
w = grep("grs_file", par_data[,1])
f = paste0(par_data[w,2] )
grs = read.table(f, header = TRUE, sep = "\t", as.is = TRUE, fill = FALSE)
rownames(grs) = paste0("ind", rownames(grs))


## Read in variable names
w = grep("variable_name", par_data[,1])
f = paste0(par_data[w,2] )
var_names = read.table(f, header = TRUE, sep = "\t", as.is = TRUE, fill = FALSE)


## Read in the QC'd metabolite data
w = grep("metabolite_qc_file", par_data[,1])
f = paste0(par_data[w,2] )
qc_metabo_data = read.table(f, header = TRUE, sep = "\t", as.is = TRUE, fill = FALSE)


## Read in the QC'd feature data
w = grep("metabolite_feature_file", par_data[,1])
f = paste0(par_data[w,2] )
qc_feature_data = read.table(f, header = TRUE, sep = "\t", as.is = TRUE, fill = FALSE)


```

## Source Data

The number of individuals in the source data file is `r nrow(sourcedata)`

The number of variables in the phenotype data file is `r ncol(sourcedata)`

## metaboprep Metabolite Data

The number of individuals in the QC'd data file is `r nrow(qc_metabo_data)`

The number of variables in the QC'd data file is `r ncol(qc_metabo_data)`



# Data Set Up

## Remove the "." in the sourcedata metabolite names

```{r}

k = sapply(colnames(sourcedata), function(i){
  substring(i, nchar(i)-1, nchar(i))
} )
w = which(k == ".1")
if(length(w)>0){
  colnames(sourcedata)[w] = sapply( colnames(sourcedata)[w], function(x){
    paste0( substring(x, 1, nchar(x)-2 ) , "_1")
  })
}
#######
#######
k = sapply(colnames(sourcedata), function(i){
  substring(i, nchar(i)-1, nchar(i))
} )
w = which(k == ".3")
if(length(w)>0){
  colnames(sourcedata)[w] = sapply( colnames(sourcedata)[w], function(x){
    paste0( substring(x, 1, nchar(x)-2 ) , "_3")
  })
}

```


## Remove individuals from source data QC'd by metaboprep

```{r}
w = which( sourcedata$ID %in% rownames(qc_metabo_data) )
sourcedata = sourcedata[w,]
```

## Remove metabolites from source data QC'd by metaboprep

```{r}
w = which( colnames(sourcedata) %in% colnames(qc_metabo_data) )
## covariate data is in columns 1:99
keep = c(1:99, w)
sourcedata = sourcedata[,keep]
```

## Update metabolite names

```{r}
m = match( qc_feature_data$ids, colnames(sourcedata)  )
colnames(sourcedata)[m] = qc_feature_data$new_id
```

## Update covariate names

```{r}
m = match( var_names$covar_names, colnames(sourcedata)  )
colnames(sourcedata)[m] = var_names$new_names
```

## Redefine the visit date variable

```{r}
## year month covariate
x = sourcedata$visit_date
x = sapply(x, function(i){ 
  o = strsplit(i, split = "\\/")[[1]]
  if(length(o)>0){
    if( nchar(o[1]) == 1 ){
      o = paste0(o[3], "-0", o[1])
    } else {
      o = paste0(o[3], "-", o[1])		
    }
  } else {o = NA}
  return(o)
})
### Redefine visitdd variable
sourcedata$visit_date = x
### insure proper conversion of NAs
w = grep("NA", sourcedata$visit_date)
if(length(w)> 0){ sourcedata$visit_date[w] = NA	}
```

## Add the GRS data

```{r}
m = match(sourcedata$ID, rownames(grs) )
mydata = cbind(sourcedata[,1:7,], grs[m,], sourcedata[, -c(1:7)] )
```

## Define the covariate classes

```{r}
for(i in 1:nrow(var_names) ){
  state = var_names$class[i]
  v = var_names$new_names[i]
  ####
  w = which( colnames(mydata) == v)
  ####
  if(state == "numerical"){
    mydata[,w] = as.numeric(mydata[,w])
  }
  if(state == "factor"){
    mydata[,w] = as.factor(mydata[,w])
  }
  if(state == "character"){
    mydata[,w] = as.character(mydata[,w])
  }
}
```


## Remove the one individual self-described as "Other" or of non-white ethnicity

```{r}
w = which(mydata$ethnicity_self == 18)
# length(w) ## 1
if(length(w) > 0){ mydata = mydata[-w, ]  }
```

- This step remove a total of `r length(w)` individuals.

## Identify covariates to remove for QC

### Define covariate vector

```{r}
covariates = var_names$new_names
length(covariates)
```

### remove study information that are not viable covariables

```{r}
i = c( "ID" , "pweight", "ethnicity_self", "ethnicity")
w = which(covariates %in% i)
length(w)
if(length(w) > 0){ covariates = covariates[ -w ]  }
length(covariates)
```

### covariate categories

```{r}
w = which(var_names$new_names %in% covariates)

cat(paste0("number of categories\n"))
length( table(var_names$category[w]) )

cat(paste0("\n\nnumber of elements in each category\n"))
table(var_names$category[w]) 
```


### remove those with fewer than 1000 observations

```{r}
## remove variables with more than 20% missingness
  # mis = apply(mydata[, covariates], 2, function(x){ sum(is.na(x))/length(x)  })
  # w = which(mis > 0.2)
  # if(length(w)>0){ covariates = covariates[-w] }

## remove variables with fewer than 1000 observations
obs_count = apply( mydata[, covariates], 2, function(x){ sum(!is.na(x)) })
w = which(obs_count < 1000)
# length(w)
if(length(w)>0){ covariates = covariates[-w] }
length(covariates)

```

- This step remove a total of `r length(w)` covariates.

### remove covariates with no variation 

```{r}
v = apply(mydata[, covariates], 2, function(x){ length(table(x))  })
w = which(v < 2)
# length(w)
if(length(w)>0){ covariates = covariates[-w] }
length(covariates)
```

- This step remove a total of `r length(w)` covariates.


### remove binary covariates with to few (n=10) observations

```{r}
v = apply(mydata[, covariates], 2, function(x){ 
  l = table(x)
  if( length(l) == 2 ){
    m = min(l)
    if( m < 10 ){
      return("remove")
    }
  } else {  
    return("keep")
  } 
})
w = which(v == "remove")
###########
if(length(w)>0){ covariates = covariates[-w] }
length(covariates)

```

- This step remove a total of `r length(w)` covariates.

## Remove those QC'd covariates from the data frame.

```{r}
temp = var_names$new_names[ !var_names$new_names %in% covariates ]
remove = temp[-c(1,2)]
###
r = which( colnames(mydata) %in% remove  )
if(length(r)>0){  mydata = mydata[,-r]  }
```

- A total of `r length(r)` covariates were removed from the data frame given the covariate QC above. 

## Redefine Sex

```{r}
w = which(mydata$sex == 1)
mydata$sex = as.character(mydata$sex)
mydata$sex[w] = "male"
mydata$sex[-w] = "female"
mydata$sex = as.factor(mydata$sex)

cat(paste0("sex redefined\n"))
```


## Redefine Subpop
```{r}
w = which(mydata$subpop == 1)
mydata$subpop = as.character(mydata$subpop)
mydata$subpop[w] = "Leiderdorp"
mydata$subpop[-w] = "Leiden"
mydata$subpop = as.factor(mydata$subpop)

cat(paste0("sub-population samples redefined\n"))

```

## Edit other variables

```{r}

## Redefine packyears
cat(paste0("redefine packyears as numeric\n"))
mydata$packyears = as.numeric(as.character(mydata$packyears))


## Redefine packyears imputed
mydata$packyears_imputed = as.numeric(as.character(mydata$packyears_imputed))


## Redefine smoking
cat(paste0("redefine smoking as numeric and as never|ever factor\n"))
mydata$smoking = as.numeric( as.character(mydata$smoking) )
# mydata$smoking = as.ordered(mydata$smoking)
mydata$smoking_ever_never = as.character( mydata$smoking )
w = which( mydata$smoking_ever_never == "0" )
mydata$smoking_ever_never[w] = "never"
mydata$smoking_ever_never[-w] = "ever"
mydata$smoking_ever_never = as.factor(mydata$smoking_ever_never)

w = which(colnames(mydata) == "smoking")
q = which(colnames(mydata) == "smoking_ever_never")

mydata = mydata[, c(1:w,q,(w+1):(q-1) )]

## Redefine education year
cat(paste0("redefine education as numeric\n"))
mydata$edu_year = as.numeric(as.character(mydata$edu_year))

## Redefine income
cat(paste0("redefine income as numeric\n"))
mydata$income = gsub("8", NA ,mydata$income)
mydata$income = gsub("9", NA ,mydata$income)
# mydata$income = as.ordered(mydata$income)
mydata$income = as.numeric(as.character(mydata$income))

## Redefine household income
cat(paste0("redefine househould income as numeric\n"))
mydata$income_hh = gsub("8", NA ,mydata$income_hh)
mydata$income_hh = gsub("9", NA ,mydata$income_hh)
# mydata$income_hh = as.ordered(mydata$income_hh)
mydata$income_hh = as.numeric(as.character(mydata$income_hh))

## Redefine currently educated
cat(paste0("redefine curently educated numeric\n"))
mydata$edu_currently = as.character(mydata$edu_currently)
mydata$edu_currently = gsub("9", NA ,mydata$edu_currently)
# mydata$edu_currently = as.ordered( factor(mydata$edu_currently, levels = c("3","2","1") ) )
w = which( mydata$edu_currently %in% c("1","2") )
mydata$edu_currently[w] = "yes"; mydata$edu_currently[-w] = "no"
mydata$edu_currently = as.factor(mydata$edu_currently)

```

## Define metabolite states

```{r}

### FASTING DATA
fast = grep("_f", colnames(mydata))
fast = colnames(mydata)[fast]
## remove unwanted
w = grep("_fibrate", fast)
if(length(w)> 0){fast = fast[-w]}

### POSTPRANDIAL DATA
post = grep("_p", colnames(mydata))
post = colnames(mydata)[post]

### RESPONSE
respo = grep("_r", colnames(mydata))
respo = colnames(mydata)[respo]

traits = c(fast, post, respo)

```

## Define metabolite IDs

```{r}
metids = gsub("_f","", fast)
```


```{r, include = FALSE}
# skimr::skim( mydata[, 1:5] )
```


## Define study_data list

```{r}
study_data = list(working_data = mydata,
                  metids = metids,
                  traits = traits,
                  fasting = fast,
                  postprandial = post,
                  response = respo,
                  source_data = sourcedata,
                  covariate_anno = var_names,
                  covariables = covariates,
                  feature_anno = qc_feature_data)

```

## Write study_data R object to file

```{r}
w = grep("data_dir", par_data[,1])
f = paste0(par_data[w,2], "processed/my_working_QCd_study_data.Rdata" )
save(study_data, file = f)
```
