==> 03a_explore_paired_qc_methods.Rmd <==
---
title: "QC of metabolite data"
author: "David Hughes"
date: "1/9/2022"
output: 
  pdf_document:
    number_sections: true
---

This PDF report will identify for each metabolite and metabolite pair (fasting & postprandial), outliers in the distribution and turn them into NAs.



```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.width = 4, fig.height = 4, echo = TRUE, error = FALSE, warning = FALSE, message = FALSE)

library(tidyverse)
library(quantreg)
library(patchwork)
```

## Load the data

```{r}
#########################
### LOAD Data
#########################
### parameter file
pfile = "parameters/pfile.txt"
par_data = read.table(pfile, header = FALSE, sep = "=", as.is = TRUE)

## load study R data object
cat(paste0("Load the study data Rdata file\n"))
w = grep("data_dir", par_data[,1])
f = paste0(par_data[w,2], "processed/my_working_QCd_study_data.Rdata" )
load(f)

## Define mydata: the working data object
cat(paste0("Define mydata working data frame\n"))
mydata = study_data$working_data

```

# Read in the data

```{r }
# source("LOAD.R")

```


# 'mydata' is my working study data set

## Let's print the fasting -v- postprandial scatter plots to file along with the identification of outliers given a variety of different steps. 

## QC steps:

1) SINGLE FEATURE|METABOLITE outliers
  - remove zero values; turn into NAs
  - inter-quartile range distances from the median, using the function id_outliers()

2) PAIRED (fasing -v- postprandial) FEATURE|METABOLITE outliers
  - methods of choice:
    - standard deviations
    - euclidean deltas
    - linear model residual distances
    - linear model cook distance

## Metabolite outlier filtering

```{r}
source("functions/paired_outlier_filtering.R")
source("functions/id_outliers.R")
```

```{r}
id = "shdll"
fasting = paste0(id, "_f")
postprandial = paste0(id, "_p")
df = paired_outlier_filtering(wdata = mydata, 
                             trait1 = fasting,
                             trait2 = postprandial,
                             IQR_distance = 5, 
                             paired_SD_distance = 5, 
                             paired_delta_distance = 5, 
                             residual_distance = 5,
                             cooks_distance = 0.025)
```

```{r}
source("functions/inspection_plot.R")
```


```{r, fig.width = 8, fig.height = 6}
# xlvldltgpct
# shdll
inspection_plot(wdata = mydata, metabolite = "shdll")
```

```{r, fig.width = 8, fig.height = 6}
inspection_plot(wdata = mydata, metabolite = "sldll")
```

```{r}
my_inspection_plots = lapply(metids, function(id){
    inspection_plot(wdata = mydata, metabolite = id)
})
```

```{r, include = FALSE}
##############################
## Write plots to a single
## pdf
##############################
w = which(par_data[,1] == "results_dir")
f = paste0( par_data[w,2] , "figures/Fasting_Postprandial_Outlier_Plots_2.pdf")
pdf(f, width = 9, height = 11)

ggpubr::ggarrange(plotlist = my_inspection_plots, 
                  ncol = 1, nrow = 2)

dev.off()
```
