==> 03b_metaboliteQC_evaluation.Rmd <==
---
title: "QC of metabolite data"
author: "David Hughes"
date: "1/9/2022"
output: 
  pdf_document:
    number_sections: true
---

This PDF report will identify for each metabolite and metabolite pair (fasting & postprandial), outliers in the distribution and turn them into NAs.



```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.width = 4, fig.height = 4, echo = TRUE, error = FALSE, warning = FALSE, message = FALSE)

library(tidyverse)
library(quantreg)
library(patchwork)
```

# Read in the data

```{r }
source("LOAD.R")
```
# 'mydata' is my working study data set

## Let's print the fasting -v- postprandial scatter plots to file along with the identification of outliers given extreme delta estimates (5 IQR from the median of postprandial - fasting). 

## QC steps:

1) SINGLE FEATURE|METABOLITE outliers
  - remove zero values; turn into NAs
  - inter-quartile range distances from the median, using the function id_outliers()

2) PAIRED (fasing -v- postprandial) FEATURE|METABOLITE outliers
  - methods of choice:
    - simple euclidean deltas
    
## Metabolite outlier filtering

```{r}
id = "shdll"
fasting = paste0(id, "_f")
postprandial = paste0(id, "_p")
df = paired_metabolite_qc(wdata = mydata,
                          trait1 = fasting,
                          trait2 = postprandial,
                          single_trait_IQR_distance = 5,
                          paired_delta_distance = 5)
```


```{r, fig.width = 10, fig.height = 4}
# xlvldltgpct
# shdll

paired_metabolite_qc_plots(wdata = mydata, metabolite = "xxlvldlp")
```

```{r, fig.width = 10, fig.height = 4}
paired_metabolite_qc_plots(wdata = mydata, metabolite = "sldll")
```

```{r}
my_paired_qc_plots = lapply(metids, function(id){
    paired_metabolite_qc_plots(wdata = mydata, metabolite = id)
})
#### add metabolite ids to list
names(my_paired_qc_plots) = metids

### save plot list to file
# w = which(par_data[,1] == "results_dir")
# f = paste0( par_data[w,2] , "my_paired_qc_plots.Rdata")
# save(my_paired_qc_plots, file = f)

```

```{r, include = FALSE}
##############################
## Write plots to a single
## pdf
##############################
w = which(par_data[,1] == "results_dir")
f = paste0( par_data[w,2] , "figures/Fasting_Postprandial_Paired_QC_Plots.pdf")
pdf(f, width = 9, height = 11)

ggpubr::ggarrange(plotlist = my_paired_qc_plots, 
                  ncol = 1, nrow = 4)

dev.off()
```



```{r, fig.width = 8, fig.height = 4}
id = "svldlfcpct"
p = paired_metabolite_qc_plots(wdata = mydata, metabolite = id)
p
```










#### OLD ####


```{r}
## Deming regression
form0 = as.formula( paste0(p , " ~ ", f) )
fitD = deming(form0, data = df)
```


## Plot linear, quadratic, cubic, GAM, and Demming regression line

```{r, fig.width = 6, fig.height = 4}
meanF = mean(df[,1])
meanP = mean(df[,2])

plot = df %>% ggplot(aes_string(x = f, y = p)) +
  ## points
  geom_point( shape = 21, fill = "grey80", alpha = 0.5, size = 4) +
  ## equivalency line
  geom_abline( aes(intercept = 0, slope = 1, 
               color = "equivalency"), 
               linetype = "dashed"  ) +
  
  ## deming
  geom_abline( aes(intercept = fitD$coefficients[1], slope = fitD$coefficients[2], 
               color = "Deming"),
               size = 1.25, linetype = "dashed") +
  
  ## mean fasting
  geom_vline( xintercept = meanF, color = "grey60", linetype = "dashed"  ) +
  ##  mean postprandial
  geom_hline( yintercept = meanP, color = "grey60", linetype = "dashed"  ) +
  
  ## linear
  geom_smooth(method = "lm", formula = y ~ x, aes(color = "linear") ) +
  
  ## polynomials
  geom_smooth(method = "lm", formula = y ~ poly(x,2), aes(color = "quadratic") ) +
  geom_smooth(method = "lm", formula = y ~ poly(x,3), aes(color = "cubic") ) +
  
  ## gam
  geom_smooth(method = "gam", formula = y ~ s(x), aes(color = "GAM") ) +
  
  theme_bw() +
  xlim(min(df[, 1:2]),max(df[, 1:2])) + 
  ylim(min(df[, 1:2]),max(df[, 1:2])) +
  scale_color_manual(name="model", 
                     values = c( "equivalency" = "black",
                                 "Deming" = "red3", 
                                 "linear" = "orangered1", 
                                 "quadratic" = "blue", 
                                 "cubic" = "purple", 
                                 "GAM" = "green"),
                     guide = "legend") 

####  
plot

```


```{r, fig.width = 10, fig.height = 6}

id = "lvldlplpct"
f = paste0(id, "_f")
p = paste0(id, "_p")
##
p1 = df %>% ggplot(aes_string(x = f, y = p)) +
  geom_point(aes(color = sd_outliers)) +
  scale_color_brewer(palette = "Set1")

p2 = df %>% ggplot(aes_string(x = f, y = p)) +
  geom_point(aes(color = delta_outliers)) +
  scale_color_brewer(palette = "Set1")

p3 = df %>% ggplot(aes_string(x = f, y = p)) +
  geom_point(aes(color = residual_outliers)) +
  scale_color_brewer(palette = "Set1")

p4 = df %>% ggplot(aes_string(x = f, y = p)) +
  geom_point(aes(color = cooks_outliers)) +
  scale_color_brewer(palette = "Set1")


(p1|p2)/(p3|p4)


```




```{r}
df = paired_outlier_filtering(wdata = mydata, 
                             fasting = "leu_f",
                             postprandial = "leu_p",
                             IQR_distance = 5, 
                             paired_SD_distance = 10, 
                             paired_delta_distance = 10, 
                             residual_distance = 4,
                             cooks_distance = 0.01)
```



```{r, fig.width = 12, fig.height = 10}

id = "leu"
f = paste0(id, "_f")
p = paste0(id, "_p")
##
p1 = df %>% ggplot(aes_string(x = f, y = p)) +
  geom_point(aes(color = sd_outliers)) +
  scale_color_brewer(palette = "Set1")

p2 = df %>% ggplot(aes_string(x = f, y = p)) +
  geom_point(aes(color = delta_outliers)) +
  scale_color_brewer(palette = "Set1")

p3 = df %>% ggplot(aes_string(x = f, y = p)) +
  geom_point(aes(color = residual_outliers)) +
  scale_color_brewer(palette = "Set1")


p4 = df %>% ggplot(aes_string(x = f, y = p)) +
  geom_point(aes(color = cooks_outliers)) +
  scale_color_brewer(palette = "Set1")


(p1|p2)/(p3|p4)

```



## a function to identify outliers

```{r}
na_outliers = function(wdata, outlier_values){
  for(i in 1:ncol(wdata)){
    neg = which( wdata[,i] < outlier_values[i,1] )
    pos = which( wdata[,i] > outlier_values[i,2] )
    outliers = c(neg, pos)
    if(length(outliers)>0){
      wdata[outliers,i] = NA
    }  
  }
  return(wdata)
}
```

```{r}
id = "leu"
# id = "idlpl"
f = paste0(id, "_f")
p = paste0(id, "_p")
## define the data set
df = mydata[, c(f, p) ]
## remove zero values
df[df == 0] = NA
## estimate interquartile range
iqr = t( apply(df, 2, function(x){ est_iqr(x, 5 ) }) )
## identify outliers
df = na_outliers(df, iqr)
## remove missing values
df = na.omit(df)

```


# Is the fasting -v- postprandial data better fit with a non-linear model?

```{r}
l_nl_test = t( sapply(metids, function(id){
  f = paste0(id, "_f")
  p = paste0(id, "_p")
  ## define the data set
  df = mydata[, c(f, p) ]
  
  ## remove zero values
  df[df == 0] = NA
    
  ## estimate interquartile range
  iqr = t( apply(df, 2, function(x){ est_iqr(x, 5 ) }) )
    
  ## identify outliers
  df = na_outliers(df, iqr)
  
  ## remove missing values
  df = na.omit(df)
    
  ## define model formulas
  form0 = as.formula( paste0(p , " ~ ", f) )
  form_gam = as.formula( paste0(p , " ~  s(", f, ")") )
  form_quadratic = as.formula( paste0(p , " ~  poly(", f, ",2)") )
  form_cubic = as.formula( paste0(p , " ~  poly(", f, ",3)") )
  
  fit0 = gam(form0, data = df)
  fitG = gam(form_gam, data = df)
  a = anova(fit0, fitG, test = "F")
  gam_p = a[2,6]
  gam_rss = a[2,2]
  
  fit0 = lm(form0, data = df)
  fitQ = lm(form_quadratic, data = df)
  fitC = lm(form_cubic, data = df)
  a = anova(fit0, fitQ, fitC, test = "F")
  quadratic_p = a[2,6]
  quandratic_rss = a[2,2]
  cubic_p = a[3,6]
  cubic_rss = a[3,2]
  
  
  #fitD = deming(form0, data = df)
  
  out = c(gam_rss, quandratic_rss, cubic_rss,
          gam_p, quadratic_p,cubic_p)
  names(out) = c("gam_rss", "quandratic_rss", "cubic_rss",
          "gam_p", "quadratic_p", "cubic_p")
  
  return(out)
  
  
}) )
##########
l_nl_test = as.data.frame(l_nl_test)
```

## Range of p-values for the GAM model

```{r}
summary(l_nl_test[, 4])
```
## How many metabolites are better fit with a GAM ?

```{r}
## count
sum(l_nl_test[, 4]<= 0.05/40) 
## Percent of total metabolites tested
sum(l_nl_test[, 4]<= 0.05/40) / nrow(l_nl_test)
```

```{r}
l_nl_test %>% arrange(gam_p) %>% kable() %>% kableExtra::kable_classic()
```


## which metabolites are NOT a better fit by a GAM ?

```{r}
## count
sum(l_nl_test[, 4] > 0.05/40) 
## names
names( which(l_nl_test[, 4] > 0.05/40) )
```

## Estimate the Demming model

```{r}
id = "lvldlplpct"
# id = "idlpl"
f = paste0(id, "_f")
p = paste0(id, "_p")
## define the data set
df = mydata[, c(f, p) ]
## remove zero values
df[df == 0] = NA
## estimate interquartile range
iqr = t( apply(df, 2, function(x){ est_iqr(x, 5 ) }) )
## identify outliers
df = na_outliers(df, iqr)
## remove missing values
df = na.omit(df)

## Demming regression
form0 = as.formula( paste0(p , " ~ ", f) )
fitD = deming(form0, data = df)
```


## Plot linear, quadratic, cubic, GAM, and Demming regression line

```{r, fig.width = 6, fig.height = 4}
meanF = mean(df[,1])
meanP = mean(df[,2])

plot = df %>% ggplot(aes_string(x = f, y = p)) +
  ## points
  geom_point( shape = 21, fill = "grey80", alpha = 0.5, size = 4) +
  ## equivalency line
  geom_abline( aes(intercept = 0, slope = 1, 
               color = "equivalency"), 
               linetype = "dashed"  ) +
  
  ## deming
  geom_abline( aes(intercept = fitD$coefficients[1], slope = fitD$coefficients[2], 
               color = "Deming"),
               size = 1.25, linetype = "dashed") +
  
  ## mean fasting
  geom_vline( xintercept = meanF, color = "grey60", linetype = "dashed"  ) +
  ##  mean postprandial
  geom_hline( yintercept = meanP, color = "grey60", linetype = "dashed"  ) +
  
  ## linear
  geom_smooth(method = "lm", formula = y ~ x, aes(color = "linear") ) +
  
  ## polynomials
  geom_smooth(method = "lm", formula = y ~ poly(x,2), aes(color = "quadratic") ) +
  geom_smooth(method = "lm", formula = y ~ poly(x,3), aes(color = "cubic") ) +
  
  ## gam
  geom_smooth(method = "gam", formula = y ~ s(x), aes(color = "GAM") ) +
  
  theme_bw() +
  xlim(0,max(df)) + 
  ylim(0,max(df)) +
  scale_color_manual(name="model", 
                     values = c( "equivalency" = "black",
                                 "Deming" = "red3", 
                                 "linear" = "orangered1", 
                                 "quadratic" = "blue", 
                                 "cubic" = "purple", 
                                 "GAM" = "green"),
                     guide = "legend") 

####  
plot

```


## which metabolites are BEST fit by a GAM ?

```{r}
names( sort(l_nl_test[, 4]) )[1:15]
```


```{r}

id = "xxlvldlp"
id = "leu"

f = paste0(id, "_f")
p = paste0(id, "_p")
## define the data set
df = mydata[, c(f, p) ]
## remove zero values
df[df == 0] = NA
## estimate interquartile range
iqr = t( apply(df, 2, function(x){ est_iqr(x, 5 ) }) )
## identify outliers
df = na_outliers(df, iqr)
## remove missing values
df = na.omit(df)

## Deming regression
form0 = as.formula( paste0(p , " ~ ", f) )
fitD = deming(form0, data = df)
```


```{r, fig.width = 10, fig.height = 6}

df %>% ggplot(aes_string(x = f, y = p)) +
  geom_point( shape = 21, fill = "grey80", alpha = 0.5, size = 4) +
  ## equivalency line
  geom_abline( intercept = 0, slope = 1, color = "grey", linetype = "dashed"  ) +
  
  ## linear
  geom_smooth(method = "lm", formula = y ~ x, color = "blue"  ) +
  
  ## polynomials
  geom_smooth(method = "lm", formula = y ~ poly(x,2), color = "purple" ) +
  geom_smooth(method = "lm", formula = y ~ poly(x,3), color = "black" ) +
  
  ## gam
  geom_smooth(method = "gam", formula = y ~ s(x), color = "green" ) +
  
  ## deming
  geom_abline( intercept = fitD$coefficients[1], slope = fitD$coefficients[2], color = "red",
               size = 1.5, linetype = "dashed") +
  
  theme_bw() +
  xlim(0,max(df)) + 
  ylim(0,max(df)) 
  

```

## Which model type has the smallest RSS

```{r}

```


```{r, fig.width = 10, fig.height = 6}

nls_fit = nls(leu_p ~ SSlogis(log(leu_f) , Asym, xmid, scal), data = df)
#####
ss = coef(summary(nls_fit))
asym_start = ss["Asym", 1]
xmid_start = ss["xmid", 1]
scal_start = ss["scal", 1]


df %>% ggplot(aes_string(x = f, y = p)) +
  geom_point( shape = 21, fill = "grey80", alpha = 0.5, size = 4) +
  ## equivalency line
  geom_abline( intercept = 0, slope = 1, color = "grey", linetype = "dashed"  ) +
  
  ## linear
  geom_smooth(method = "lm", formula = y ~ x, color = "orange"  ) +
  
  ## polynomials
  geom_smooth(method = "lm", formula = y ~ poly(x,2), color = "dodgerblue" , se = FALSE) +
  geom_smooth(method = "lm", formula = y ~ poly(x,3),  color = "royalblue", se = FALSE ) +
  
  ## NLS
  geom_smooth(method = "nls",
              formula = y ~ Asym/(1 + exp((xmid - log(x))/scal)), 
              method.args = list(start = list(Asym = asym_start, xmid = xmid_start, scal = scal_start)),
              #data = DNase1,
              se = FALSE,
              color = "black") + 
  
  ## gam
  geom_smooth(method = "gam", formula = y ~ s(x), color = "green2", se = FALSE ) +
  
  ## deming
  geom_abline( intercept = fitD$coefficients[1], slope = fitD$coefficients[2], color = "red",
               size = 1.5, linetype = "dashed") +
  
  theme_bw() +
  xlim(0,max(df)) + 
  ylim(0,max(df)) 

```

```{r, fig.width = 6, fig.height = 4}
df2 = df[, 1:2]
df2[,1] = scale( df[,1] )
df2[,2] = scale( df[,2] )

##  non linear least squares
nls_fit = nls( xxlvldlp_p ~ SSlogis( log(xxlvldlp_f) , Asym, xmid, scal), data = df)
ss = coef(summary(nls_fit))
asym_start = ss["Asym", 1]
xmid_start = ss["xmid", 1]
scal_start = ss["scal", 1]

df2$fitted = scale( fitted(nls_fit) )

## Scale the values
xySD = apply(df2, 1, function(z){ sd(z, na.rm = TRUE) })
SD = mean(xySD)

## Define the linear outliers
df2$class = apply(df2[,1:2], 1, function(z){
        ifelse( abs(z[2]-z[1]) <= SD*8, "good" , "outlier") 
        })

##
df2$class2 = apply(df2[, 1:3], 1, function(x){
  ifelse( abs( x[3] - x[1] ) >= SD | abs( x[3] - x[1] ) >= SD , "good", "outlier" )
})


df2 %>% ggplot(aes(x = xxlvldlp_f, y = xxlvldlp_p)) +
  geom_point( aes(color = class) ) +
  geom_smooth() +
  theme_bw()

      
```

## ONLS fitting

```{r}
s = sample(1:nrow(df2), 600)
df3 = df[s,1:2 ]
onls_fit = onls(xxlvldlp_p ~ Asym/(1 + exp((xmid - log(xxlvldlp_f))/scal)),
             data = df3, 
             start = list(Asym = asym_start, xmid = xmid_start, scal = scal_start) )


```


```{r, fig.width = 10, fig.height = 6}

df3 %>% ggplot(aes_string(x = f, y = p)) +
  geom_point( shape = 21, fill = "grey80", alpha = 0.5, size = 4) +
  ## equivalency line
  geom_abline( intercept = 0, slope = 1, color = "grey", linetype = "dashed"  ) +
  
  ## linear
  geom_smooth(method = "lm", formula = y ~ x, color = "orange"  ) +
  
  ## polynomials
  geom_smooth(method = "lm", formula = y ~ poly(x,2), color = "dodgerblue" , se = FALSE) +
  geom_smooth(method = "lm", formula = y ~ poly(x,3),  color = "royalblue", se = FALSE ) +
  
  ## NLS
  # geom_smooth(method = "nls",
  #             formula = y ~ Asym/(1 + exp((xmid - log(x))/scal)),
  #             method.args = list(start = list(Asym = asym_start, xmid = xmid_start, scal = scal_start)),
  #             #data = DNase1,
  #             se = FALSE,
  #             color = "black") +
  # 
  ## NLS
  geom_smooth(method = "onls",
              formula = y ~ Asym/(1 + exp((xmid - log(x))/scal)), 
              method.args = list(start = list(Asym = asym_start, xmid = xmid_start, scal = scal_start)),
              #data = DNase1,
              se = FALSE,
              color = "black") +

  ## gam
  # geom_smooth(method = "gam", formula = y ~ s(x), color = "green2", se = FALSE ) +
  
  ## deming
  # geom_abline( intercept = fitD$coefficients[1], slope = fitD$coefficients[2], color = "red",
  #              size = 1.5, linetype = "dashed") +
  
  theme_bw() +
  xlim(0,max(df3)) + 
  ylim(0,max(df3)) 

```


```{r, fig.width=10, fig.height = 5}
library(patchwork)
###
fit = lm(xxlvldlp_p ~ xxlvldlp_f, data = df3)
fitq = lm(xxlvldlp_p ~ poly(xxlvldlp_f,2), data = df3)
exdata = data.frame(onls_res = onls_fit$residONLS, 
                    onls_res2 = onls_fit$resid_o,
                    nls_res = onls_fit$residNLS, 
                    lm_res = residuals(fit),
                    quadratic_res = residuals(fitq))
####
p1 = exdata %>% ggplot(aes(x = lm_res, y = nls_res)) + 
  geom_point()
###
p2 = exdata %>% ggplot(aes(x = lm_res, y = quadratic_res)) + 
  geom_point()
###
p3 = exdata %>% ggplot(aes(x = nls_res, y = quadratic_res)) + 
  geom_point()
###
p4 = exdata %>% ggplot(aes(x = nls_res, y = onls_res2)) + 
  geom_point()


( p1 | p2 ) / (p3 | p4)



```


```{r}

lm_fit = lm(glc_p ~ glc_f, data = df)
###
nls_fit = nls(glc_p ~ SSlogis(log(glc_f) , Asym, xmid, scal), data = df)
###
df2 = df
df2[,1] = moosefun::ztransform( df[,1] )
df2[,2] = moosefun::ztransform( df[,2] )


df2 = df
df2 = apply(df2, 2, rntransform)
df2 = as.data.frame(df2)
onls_fit = onls(glc_p ~ SSlogis(log(glc_f) , Asym, xmid, scal),
             data = df2, 
             start = list(Asym = asym_start, xmid = xmid_start, scal = scal_start) )

anova(lm_fit, nls_fit)

###
ss = coef(summary(nls_fit))
asym_start = ss["Asym", 1]
xmid_start = ss["xmid", 1]
scal_start = ss["scal", 1]

###
df = na.omit(df)
fit0 = lm(glc_p ~ glc_f, data = df)
fitp = lm(glc_p ~ poly(glc_f,2), data = df)
anova(fit0, fitp)


```

```{r}


df %>% ggplot(aes_string(x = f, y = p)) +
  geom_point( shape = 21, fill = "grey30", alpha = 0.5, size = 4) +
  # geom_smooth(method = "gam", formula = y ~ s(x), color = "green" ) +
  geom_smooth(method = "lm", formula = y ~ x ) +
  geom_abline( intercept = fitD$coefficients[1], slope = fitD$coefficients[2], color = "green"  ) +
  # geom_smooth(method = "lm", formula = y ~ poly(x,3), color = "green" ) +
  geom_smooth(method = "nls", 
              formula = y ~ SSlogis(log(x), Asym, xmid, scal) , 
              data = df,
              se = FALSE,
              color = "red") +
  geom_smooth(method = "onls",
              formula = y ~ Asym/(1 + exp((xmid - log(x))/scal)), 
              method.args = list( start = list(Asym = asym_start, xmid = xmid_start, scal = scal_start) ), 
              data = df,
              se = FALSE,
              color = "purple") +
  theme_bw() 
  #xlim(0,20) + 
  #ylim(0,20)

```


```{r}
new_response_traits = c()

for(id in metids[ 168:length(metids) ] ){
  cat(paste0("Now processing metabolite ", id, "\n"))
  fid = paste0(id, "_f")
  pid = paste0(id, "_p")
  temp_data = data.frame(mydata[, c(fid, pid)])
  rownames(temp_data) = mydata$ID
  ## If data for both fasting and postprandial
  ##  data is present proceed.
  if(ncol(temp_data) == 2){
    ## remove zero values
    temp_data[temp_data == 0] = NA
    
    ## estimate interquartile range
    iqr = t( apply(temp_data, 2, function(x){ est_iqr(x, 5 ) }) )
    
    ## identify outliers
    temp_data = na_outliers(temp_data, iqr)
    
    ## define the model formulas
    form = formula( paste0(pid ," ~ ", fid) )
    
    ## Deming Regression
    dfit = deming(form, data = temp_data)
    dem_res = residuals(dfit)
    m = match(rownames(temp_data), names(dem_res))
    dem_res = dem_res[m]
    
    ## Return data
    new_response_traits = cbind(new_response_traits, dem_res)
  } else {
    new_response_traits = cbind(new_response_traits, rep(NA, nrow(mydata)) )  
  }
}


```


```{r}
colnames(new_response_traits) = paste0(metids, "_deming")
```


```{r}
study_data$new_deming_response_traits = new_response_traits
```


## Write study_data R object to file

```{r}
w = grep("data_dir", par_data[,1])
f = paste0(par_data[w,2], "processed/my_working_QCd_study_data_v3.Rdata" )
save(study_data, file = f)
```
