==> 04_build_deming_response.Rmd <==
---
title: "generate a new response trait"
author: "David Hughes"
date: "11/12/2022"
output: 
  pdf_document:
    number_sections: true
---

This Rmd file will estimate a new response traits. 

1. A Deming regression residual.

---

```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.width = 4, fig.height = 4, echo = TRUE, error = FALSE, warning = FALSE, message = FALSE)

library(deming)
library(tidyverse)
# library(mgcv)
```


## Load the data

```{r}
#########################
### LOAD Data
#########################
### parameter file
pfile = "parameters/pfile.txt"
par_data = read.table(pfile, header = FALSE, sep = "=", as.is = TRUE)

## load study R data object
cat(paste0("Load the study data Rdata file\n"))
w = grep("data_dir", par_data[,1])
f = paste0(par_data[w,2], "processed/my_working_QCd_study_data_v1.Rdata" )
load(f)

## Define mydata and metids
cat(paste0("Define mydata working data frame\n"))
mydata = study_data$working_data
metids = study_data$metids

```

# 'mydata' is my working study data set

## Build the new response traits

```{r}
new_response_traits = c()
deming_intercept_slope = matrix(NA, nrow = length(metids), ncol = 2, 
                                dimnames = list(metids, c("intercept", "slope")))

for( metid in metids ){
  cat(paste0("Now processing metabolite ", metid, "\n"))
  fid = paste0(metid, "_f")
  pid = paste0(metid, "_p")
  temp_data = data.frame(mydata[, c(fid, pid)])
  rownames(temp_data) = mydata$ID
  ## If data for both fasting and postprandial
  ##  data is present proceed.
  if(ncol(temp_data) == 2){
    
    ## define the model formulas
    form = formula( paste0(pid ," ~ ", fid) )
    
    ## Deming Regression
    dfit = deming(form, data = temp_data)
    ## add intercept and slope to data frame
    is = dfit$coefficients; names(is) = c("intercept","slope")
    deming_intercept_slope[metid, ] = is
    ## extract residuals
    dem_res = residuals(dfit)
    m = match(rownames(temp_data), names(dem_res))
    dem_res = dem_res[m]
    
    ## Return data
    new_response_traits = cbind(new_response_traits, dem_res)
  } else {
    new_response_traits = cbind(new_response_traits, rep(NA, nrow(mydata)) )  
  }
}

```


```{r}
colnames(new_response_traits) = paste0(metids, "_deming_r")
```


## Add the deming data to working_data data frame to the R obj study_data

```{r}
study_data$working_data = cbind(mydata, new_response_traits)
```

## Define a full data data frame that has the fasting, postprandial, oNLS response, and deming response.

```{r}
study_data$full_data = study_data$working_data
```

## remove the oNLS response traits

```{r}
mydata = study_data$working_data

r = grep( "_r", colnames(mydata)  )
d = grep( "_deming_r", colnames(mydata)  )
w = r[!r %in% d]

mydata = mydata[, -w]
```

## Rename the _deming_r columns simply as _r

```{r}
w = grep("_deming_r" , colnames(mydata) )
##
n = colnames(mydata)[w]
n = gsub("_deming_r","_r",n)
##
colnames(mydata)[w] = n
```

## Redefine the list of traits

```{r}
metids = study_data$metids
##
fasting = paste0(metids, "_f")
postprandial = paste0(metids, "_p")
response = paste0(metids, "_r")

traits = c(fasting, postprandial, response)
```

```{r}
study_data$working_data = mydata
study_data$traits = traits
study_data$response = response
```


## Write study_data R object to file

```{r}
w = grep("data_dir", par_data[,1])
f = paste0(par_data[w,2], "processed/my_working_QCd_study_data_v2.Rdata" )
save(study_data, file = f)
```

## Write the intercept and slope data for deming regressions to file

```{r}
w = grep("results_dir", par_data[,1])
f = paste0(par_data[w,2], "Deming_regression_slope_intercept.txt" )
###
write.table(deming_intercept_slope, file = f, 
            row.names = TRUE, col.names = TRUE, 
            sep = "\t", quote = FALSE)

```
