==> 05_iPVs.Rmd <==
---
title: "estimate the number of effective markers"
author: "David Hughes"
date: "11/12/2022"
output: 
  pdf_document:
    number_sections: true
---

This Rmd file will estimate the number of effective markers

---

```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.width = 4, fig.height = 4, echo = TRUE, error = FALSE, warning = FALSE, message = FALSE)

library(iPVs)
library(tidyverse)

```


## Load the data

```{r}
#########################
### LOAD Data
#########################
### parameter file
pfile = "parameters/pfile.txt"
par_data = read.table(pfile, header = FALSE, sep = "=", as.is = TRUE)

## load study R data object
cat(paste0("Load the study data Rdata file\n"))
w = grep("data_dir", par_data[,1])
f = paste0(par_data[w,2], "processed/my_working_QCd_study_data_v2.Rdata" )
load(f)

## Define mydata and metids
cat(paste0("Define mydata working data frame\n"))
mydata = study_data$working_data
metids = study_data$metids
traits = study_data$traits
```

# 'mydata' is my working study data set

## Run iPVs

```{r}
mypvs = iPVs(mydata[, traits ], 
	cor_method = "spearman",
	dist_method = "R",
	hclust_meth = "complete",
	cutheight  = 0.5
	)
```


```{r}
dim(mypvs$iPV_table)
```

There are 43 

```{r, fig.width = 15, fig.height = 5}
## load a needed R package
library(dendextend)

## extract the IDs for your PVs
pv_ids = as.character(mypvs$iPV_table$PVs )

## define your tree as a dendrogram
dend = mypvs$workingdata$tree %>% as.dendrogram

## create a vector of colors to color your tree labels
n = labels(dend)
pcol = rep("black", length(n))
w = which(n %in% pv_ids ); pcol[w] = "medium blue"

## redefine elements of dendrogram
dend = dend %>% 
set("labels_cex", 0.25) %>% 
set("labels_col", pcol) %>% 
set("branches_k_color",  value = pcol)

## plot the dendrogram
dend %>% plot(main = paste0( "-- Principle Variables --" ) )
abline(h = 0.5, col = "red", lwd = 3)
```

```{r}
cat(paste0("Write pdf of hclust to file\n"))
w = grep("results_dir", par_data[,1])
f = paste0(par_data[w,2], "figures/hclust_dend.pdf" )

pdf(file = f, width = 20, height = 5)
dend %>% plot(main = paste0( "-- NEO: fasting, postprandial, and response metabolite trait dendrogram --" ) )
abline(h = 0.5, col = "red", lwd = 3)
dev.off()

```
