==> 06_NEO_description.Rmd <==
---
title: "NEO BMI-Metabolite Observational Analyses"
author: "David Hughes"
date: "2/6/2021"
output: 
  pdf_document:
    number_sections: true
---

This PDF report will contain all observational analysis for the NEO BMI-2-metabolite one-sample MR study. 


```{r setup, include=FALSE}
knitr::opts_chunk$set( echo = FALSE, error = FALSE, warning = FALSE, message = FALSE)

source("LOAD.R")

```



The QC'd data set contains

1. `r nrow(mydata)` individuals
2. `r ncol(mydata)` variables
  - including `r length(traits)` metabolites
  - including `r length(covariates)` covariates
  
___ 


# Population Descriptions

  1. all samples
  2. the Leiden - high BMI oversampling
  3. the Leiderdorp - random village sampling
  4. men
  5. women

## single use functions

```{r}
## sex
# male_female_count = c( sum( mydata$sex  == 1), sum( mydata$sex  == 2) )
##############################
## FUNCTION for Male Female counts
##############################
MF_count = function(x){
  count = c( sum( x  == "male"), sum( x  == "female") )
  out = paste0("M:", count[1], " F:", count[2])
  names(out) = "MF_count"
  return(out)
}
##############################
## Smokers
##############################
Smoker_count = function(x){
  count = table(x)
  out = paste0("Never:", 
               format( round( count[1]/sum(count)* 100 , digits = 2) , nsmall = 2), " Former:", 
               format( round( count[2]/sum(count)* 100 , digits = 2) , nsmall = 2), " Current:", 
               format( round( count[3]/sum(count)* 100 , digits = 2) , nsmall = 2) )
  names(out) = "Smoker_Ratio"
  return(out)
}
##############################
## FUNCTION for higher education
##############################
Ed_count = function(x){
  count = table(x)
  out = paste0("No:", count[1], " yes:", count[2])
  out = format( round( count[2]/sum(count) * 100 , digits = 2) , nsmall = 2)
  names(out) = "HigherEdRatio"
  return(out)
}
##############################
## FUNCTION TO PROCESS VALUES
##############################
report_est = function(x){
  v = c( format( round( mean( x, na.rm = TRUE ), digits = 2 ) , nsmall = 2), 
         format( round( quantile(x, probs = c(0.5, 0.025, 0.975), na.rm = TRUE ), digits = 2 ) , nsmall = 2) , 
         format( round( quantile( x, probs = c(0,1), na.rm = TRUE ), digits = 2 ) , nsmall = 2),
         format( round( var(x, na.rm = TRUE), digits = 2 ) , nsmall = 2), 
         format( round( sd(x, na.rm = TRUE) ), digits = 2 ) , nsmall = 2)
  # v = round(v, d = 2)
  names(v) = c("mean","median","lower95","upper95","min", "max", "variance", "sd")
  ##
  r = paste0("mean: ", v[1], " (95CI:",v[3], "-", v[4], ")", " (sd:",v[8],") (Var:",v[7],") " ,"(MinMax:", v[5], "-", v[6], ")" )
  names(r) = "fullreport"
  ##
  r2 = paste0( v[1], " (",v[3], "-", v[4], ")" )
  names(r2) = "report"
  
  return(list(brief_report = r2,
              full_report = r,
              all = v))
}
##############################
## FUNCTION TO Estimate Variance 
## Explained by sub-population
##############################
vexp = function(trait){
  form = formula(paste0(trait ,"~ subpop ") )
  temp = na.omit( mydata[, c(trait, "subpop")] )
  if( class(temp[,trait]) == "factor" ){
    fit = glm(form, data = temp, family = binomial() )  
  } else {
    fit = glm(form, data = temp, family = gaussian() ) 
  }
  ###
  total_ss = summary(fit)$null.deviance
  res_ss = summary(fit)$deviance
  ss_subpop = total_ss - res_ss
  varexp = ss_subpop / total_ss
  P = summary(fit)$coefficients[2,4]
  ####
  out = c(varexp, P)
  names(out) = c("var_exp","P")
  out
}

```


```{r}

###############################
## Variables to summarize 
###############################
v2sum = c("age","bmi",
          "Yengo_Weighted_GRS", 
          "height", "weight", 
          "waistC","hipC",
          "packyears", "alcohol")
 
## ALL SAMPLE
allsum = apply( mydata[, v2sum], 2, function(x){ unlist( report_est(x)[1] ) })
allMF = MF_count(mydata$sex)
allsmoke = Smoker_count(mydata$smoking)
alled = Ed_count(mydata$edu_higher)
NEO = c(allMF, allsum, allsmoke, alled)

## MALE
w = which(mydata$sex == "male")
malesum = apply(mydata[w, v2sum], 2, function(x){  unlist( report_est(x)[1] ) })
maleMF = MF_count(mydata$sex[w])
malesmoke = Smoker_count(mydata$smoking[w])
maleed = Ed_count(mydata$edu_higher[w])
Men = c(maleMF, malesum, malesmoke, maleed)

## FEMALE
w = which(mydata$sex == "female")
femalesum = apply(mydata[w, v2sum], 2, function(x){ unlist(  report_est(x)[1] ) })
femaleMF = MF_count(mydata$sex[w])
femalesmoke = Smoker_count(mydata$smoking[w])
femaleed = Ed_count(mydata$edu_higher[w])
Women = c(femaleMF, femalesum, femalesmoke, femaleed)

## Leiderdorp
w = which(mydata$subpop == "Leiderdorp")
leidersum = apply(mydata[w, v2sum], 2, function(x){ unlist(  report_est(x)[1] ) })
leiderMF = MF_count(mydata$sex[w])
leidersmoke = Smoker_count(mydata$smoking[w])
leidered = Ed_count(mydata$edu_higher[w])
Leiderdorp = c(leiderMF, leidersum, leidersmoke, leidered)

## Leiden
w = which(mydata$subpop == "Leiden")
leidensum = apply(mydata[w, v2sum], 2, function(x){ unlist(  report_est(x)[1] ) })
leidenMF = MF_count(mydata$sex[w])
leidensmoke = Smoker_count(mydata$smoking[w])
leidened = Ed_count(mydata$edu_higher[w])
Leiden = c(leidenMF, leidensum, leidensmoke, leidened)

popsums = rbind(NEO, Men, Women, Leiderdorp,  Leiden)
colnames(popsums)[c(3:4,11)] = c("BMI","BMI-PGS","smoking")

## Variation among subpopulations
VarExp  = t( sapply(c("sex", v2sum, "smoking","edu_higher"), function(x){ vexp(x) }) )
VarExp = as.data.frame(VarExp)
VarExp[,1] = paste0( format( VarExp[,1]*100 , digits = 2) , " %")
VarExp[,2] = format( VarExp[,2] , scientific = TRUE, digits = 3)

popsums = rbind( popsums, t(VarExp) )
rownames(popsums)[6:7] = c("var exp by sub-pop", "p-value")

```

```{r}

## Print table to file
w = which(par_data[,1] == "results_dir")
f = paste0( par_data[w,2] , "tables/1.Population_Summary_Stats_v1.csv")
write.table(popsums, file = f, row.names = TRUE, col.names = TRUE, sep = ",", quote  = TRUE)

```


```{r, echo = FALSE, fig.width = 12, fig.height = 5}
t(popsums) %>%
  kbl(caption = "Population Summary Statistics") %>%
  #kable_paper("hover", full_width = F)
  kable_classic(full_width = F, html_font = "Cambria")

```


```{r}
tab = mydata %>% group_by(subpop, sex ) %>% summarise(report_est(bmi)[[2]])
colnames(tab) = c("sub-population","sex","BMI summary stats")

tab %>%
  kbl(caption = "Population Summary Statistics") %>%
  #kable_paper("hover", full_width = F)
  kable_classic(full_width = F, html_font = "Cambria")

```

## How many of the metabolite have mean differences by sub-population?

```{r}
Metabilite_VarExp_by_subpop  = t( sapply(traits, function(x){ vexp(x) }) )
cat(paste0("The integer count of metabolites that differ among sub-populations?\n"))
sum(Metabilite_VarExp_by_subpop[, 2]<0.05/43)
cat(paste0("The proportion of metabolites that differ among sub-populations?\n"))
sum(Metabilite_VarExp_by_subpop[, 2]<0.05/43) / nrow(Metabilite_VarExp_by_subpop)

```

## Distribution of Deviance Explained by sub-population

```{r}
hist(Metabilite_VarExp_by_subpop[, 1], xlab = "deviance explained", 
     main = paste0("metabolite variation explained by sub-population") )
```




## W stats and % of data "normal"

```{r}
W = apply(mydata[, traits], 2, function(x){  
  x = na.omit(x)
  if(length(x) > 5000){
    shapiro.test(  sample(x, 5000)   )$stat
  } else {
      shapiro.test(  x  )$stat
  }
  
  })
summary(W)
sum(W<0.95, na.rm = TRUE) / length(W)
```


## W stats and % of data "normal" after log transformation

```{r}
i = 0
W = apply(mydata[, traits], 2, function(x){  
  x = na.omit(x)
  x = log( x + (abs(min(x))*2) )
  ###
  if( length(x) < 50){
      NA
    } else { 
      if( length(x) > 5000){
        shapiro.test(  sample(x, 5000)   )$stat
      } else {
        shapiro.test(  x  )$stat
      }
  }
})

summary(W)
sum(W<0.95, na.rm = TRUE) / length(W)
```

## Varianced explained in BMI and BMI-PGS

### Person's r, squared

```{r}
cat(paste0("Variance explained in BMI by GRS in Leiderdorp\n"))
w = which(mydata$subpop == "Leiderdorp")
(cor.test(mydata$bmi[w], mydata$Yengo_Weighted_GRS[w])$estimate)^2

cat(paste0("\nVariance explained in BMI by GRS in NEO\n"))
(cor.test(mydata$bmi, mydata$Yengo_Weighted_GRS)$estimate)^2

cat(paste0("\nVariance explained in BMI by GRS in Leiden\n"))
w = which(mydata$subpop == "Leiden")
(cor.test(mydata$bmi[w], mydata$Yengo_Weighted_GRS[w])$estimate)^2
```

### Linear model Rsquared

```{r}
cat(paste0("weighted NEO \n"))
fit = lm(bmi ~ Yengo_Weighted_GRS, weights = pweight, data = mydata)
summary(fit)$r.squared

cat(paste0("\nun-weight NEO \n"))
fit = lm(bmi ~ Yengo_Weighted_GRS, data = mydata)
summary(fit)$r.squared

cat(paste0("\nLeiderdorp \n"))
w = which(mydata$subpop == "Leiderdorp")
fit = lm(bmi ~ Yengo_Weighted_GRS, data = mydata[w,])
summary(fit)$r.squared

cat(paste0("\nLeiden \n"))
w = which(mydata$subpop == "Leiden")
fit = lm(bmi ~ Yengo_Weighted_GRS, data = mydata[w,])
summary(fit)$r.squared
```


## Variance in BMI across subgroups

```{r}
## ALL SAMPLE
NEO = c( mean(mydata$bmi), var(mydata$bmi), sd(mydata$bmi),  nrow(mydata) )
## MALE
w = which( mydata$sex == "male")
male = c( mean(mydata$bmi[w]), var(mydata$bmi[w]), sd(mydata$bmi[w]),  length(w) )
## FEMALE
w = which(mydata$sex == "female")
female = c( mean(mydata$bmi[w]),var(mydata$bmi[w]), sd(mydata$bmi[w]),  length(w) )
## Leiderdorp
w = which(mydata$subpop == "Leiderdorp")
LeiderdropSample = w
Leiderdorp = c( mean(mydata$bmi[w]),var(mydata$bmi[w]), sd(mydata$bmi[w]),  length(w)  )
## High BMI Oversampling
w = which(mydata$subpop == "Leiden")
LeidenSample = w
Leiden = c( mean(mydata$bmi[w]), var(mydata$bmi[w]), sd(mydata$bmi[w]), length(w) )



out = rbind(Leiderdorp, NEO, male, female,  Leiden )
colnames(out) = c("mean","variance","sd", "n" )

grid.arrange(tableGrob(round(out, d = 2)))


```

## Testing if there is a difference in the BMI distribution between Leiden and Liederdorp

```{r}
cat(paste0("T-test pvalue\n"))
t.test( mydata$bmi[LeidenSample], mydata$bmi[LeiderdropSample] )$p.value

cat(paste0("\nKS-test pvalue\n"))
ks.test( mydata$bmi[LeidenSample], mydata$bmi[LeiderdropSample] )$p.value
```



# Data Distributions

## Distribution of BMI and BMI-GRS

```{r bmi_GRS_distributions, fig.width = 8, fig.height = 5, warning=FALSE, error=FALSE, message=FALSE}

pcol = RColorBrewer::brewer.pal(6, "Set1")

## BMI distribution
a = shapiro.test( sample(mydata$bmi, 5000) )
p2 = mydata %>% ggplot(aes(bmi)) + 
  geom_histogram(fill = pcol[3], color = "grey") +
  geom_vline(xintercept = mean(mydata$bmi, na.rm = TRUE), color = "grey", size = 1.5 ) +
  labs(x = "BMI", title = "BMI Distribution", subtitle = paste0( "W-stat = ", signif(a$sta, d = 3) ))

## RNT BMI distribution
a = shapiro.test( sample(rntransform( mydata$bmi ), 5000) )  
p3 = mydata %>% ggplot(aes( rntransform( bmi ) )) + 
  geom_histogram(fill = pcol[4], color = "grey") +
  labs(x = "rank normal transformed BMI", title = paste0("RNTransformed BMI\nDistribution"), subtitle = paste0( "W-stat = ", signif(a$sta, d = 3) ))


## LOG BMI distribution
a = shapiro.test( sample(log( mydata$bmi ), 5000) )  
p5 = mydata %>% ggplot(aes( log( bmi ) )) + 
  geom_histogram(fill = pcol[2], color = "grey") +
  labs(x = "log transformed BMI", title = paste0("log transformed BMI\nDistribution"), subtitle = paste0( "W-stat = ", signif(a$sta, d = 3) ))

## WEIGHTED YENGO GRS
a = shapiro.test( sample( mydata$Yengo_Weighted_GRS , 5000) )  
pWY = mydata %>% ggplot(aes( Yengo_Weighted_GRS )) + 
  geom_histogram(fill = pcol[5], color = "grey") +
  geom_vline(xintercept = mean(mydata$Yengo_Weighted_GRS, na.rm = TRUE), color = "grey", size = 1.5 ) +
  labs(x = "Yengo Weighted GRS", title = paste0("Yengo Weighted BMI-GRS\nDistribution"), subtitle = paste0( "W-stat = ", signif(a$sta, d = 3) ))

## UN-WEIGHTED YENGO GRS
a = shapiro.test( sample( mydata$Yengo_Un_Weighted_GRS , 5000) )  
puWY = mydata %>% ggplot(aes( Yengo_Un_Weighted_GRS )) + 
  geom_histogram(fill = pcol[6], color = "grey") +
  geom_vline(xintercept = mean(mydata$Yengo_Un_Weighted_GRS, na.rm = TRUE), color = "grey", size = 1.5 ) +
  labs(x = "Yengo Un-weighted GRS", title = paste0("Yengo Un-Weighted BMI-GRS\nDistribution"), subtitle = paste0( "W-stat = ", signif(a$sta, d = 3) ))


## Print figure to file
# w = which(par_data[,1] == "results_dir")
# f = paste0( par_data[w,2] , "figures/1.BMI_BMIGRS_distributions.pdf")
#   pdf(f, width = 8, height = 6)
#   grid.arrange(p2, p5, p3, pWY, puWY, nrow = 2)
#   dev.off()


grid.arrange(p2, p5, p3, pWY, puWY, nrow = 2)



```

## Significant difference among sub-populations in covariates

```{r}
w = which( mydata$subpop == "Leiderdorp" )
###
x = colnames(mydata)[c(6:25)]
tests = t( sapply(x, function(i){
  a = wilcox.test( as.numeric(mydata[w, i]), as.numeric(mydata[-w, i])  )
  out = c(a$statistic, a$p.value)
  return(out)
}) )
colnames(tests) = c("Wilcox stat","pval")
tests = as.data.frame(tests)

# t.test( mydata$Yengo_Weighted_GRS[w], mydata$Yengo_Weighted_GRS[-w]  )$p.value

# grid.arrange(tableGrob(round(tests, d = 10)))

tests %>% kable() %>% kableExtra::kable_classic()
```


## Distribution of BMI in the High BMI sample and the Leiderdorp sample.


```{r, fig.width = 8, fig.height = 6}
pcol = RColorBrewer::brewer.pal(5, "Set1")


w = which( mydata$subpop == "Leiderdorp")
a = shapiro.test( mydata$bmi[w] )
p1 = mydata[w,] %>% ggplot(aes(bmi)) + 
  geom_histogram(fill = pcol[1], color = "grey" ,  binwidth = 1) +
  geom_vline(xintercept = mean(mydata$bmi[w], na.rm = TRUE), color = "grey", size = 2 ) +
  labs(x = "BMI", title = "Leiderdorp BMI Distribution", subtitle = paste0( "W-stat = ", signif(a$sta, d = 3), "; N = ", length(w) ))
##

w = which(mydata$subpop == "Leiden")
a = shapiro.test( mydata$bmi[w] )
p2 = mydata[w,] %>% ggplot(aes(bmi)) + 
  geom_histogram(fill = pcol[2], color = "grey", binwidth = 1 ) +
  geom_vline(xintercept = mean(mydata$bmi[w], na.rm = TRUE), color = "grey", size = 2 ) +
  labs(x = "BMI", title = "Leiden BMI Distribution", subtitle = paste0( "W-stat = ", signif(a$sta, d = 3), "; N = ", length(w) ))
##
##########################
## Density Ridges
##
##########################
pcol = RColorBrewer::brewer.pal(8, "Set1")[3:1]
pcol = wesanderson::wes_palette("Darjeeling1", 3, type = c("discrete") )
#pcol = wesanderson::wes_palette("BottleRocket2", 3, type = c("discrete") )

w = which( !is.na( mydata$subpop) )

ridge_plot = mydata[w, ] %>% ggplot( aes(x = bmi, y = as.factor(subpop), fill = stat(x) ) ) + 
  # geom_density_ridges(alpha=0.8, scale = 2) +
  geom_density_ridges_gradient(scale = 2, rel_min_height = 0.0001) +
  #scale_fill_viridis_c( option = "city", begin = 0.3, end = 1) + 
  scale_fill_gradient2(low = pcol[1], mid = pcol[2], high = pcol[3], midpoint = 30) +
  labs( fill = "BMI",  x = "BMI", y = "Population Sample", title = "Variation in BMI by sub-population",
    subtitle = "Leiderdorp and Leiden")


mat = matrix(c(1,2,3,3), nrow = 2, byrow = TRUE)
 
## Print figure to file
# w = which(par_data[,1] == "results_dir")
# f = paste0( par_data[w,2] , "figures/2.BMI_distributions_Leiderdrop_High.pdf")
# pdf(f, width = 8, height = 6)
# grid.arrange(p1, p2, ridge_plot, layout_matrix = mat  )
# dev.off()

grid.arrange(p1, p2, ridge_plot, layout_matrix = mat  )



```


```{r, fig.width = 7, fig.height = 4}
ridge_plot
```

```{r, echo = FALSE}
# pdata = data.frame( population = c("NEO","highBMI","Leiderdorp"),  percentage = c(1, 0.745, 0.2548), characteristic = rep("sample size", 3))
# 
# pdata$population = factor(pdata$population, levels = c("NEO","highBMI","Leiderdorp") )
# 
# ggplot(pdata, aes(fill=population, y=percentage, x=characteristic)) + 
#     geom_bar(position="dodge", stat="identity")

```

```{r ,echo = FALSE}
# mydata %>% ggplot(aes(x = subpop, y = bmi)) + 
#   #geom_density_ridges_gradient(scale = 2, rel_min_height = 0.0001) +
#   geom_violin( aes(fill = subpop)) + 
#   geom_boxplot(width=0.1) + 
#   scale_fill_manual(values = RColorBrewer::brewer.pal(3, "Set1")) +
#   labs(y = "BMI", x = "sub-population", fill = "sub-population") +
#   theme(legend.position="bottom") +
#   coord_flip() 
  
```

```{r, echo = FALSE, fig.width = 4, fig.height = 8}
# pcol = wesanderson::wes_palette("Darjeeling1", 3, type = c("discrete") )
# pcol = c("#edf8b1","#7fcdbb","#2c7fb8")
# p1 = mydata %>% ggplot(aes(y = subpop, x = bmi, fill = stat(x))) + 
#   geom_density_ridges_gradient(scale = 2, rel_min_height = 0.0001) +
#   scale_fill_gradient2(low = pcol[1], mid = pcol[2], high = pcol[3], midpoint = 30) +
#   labs(x = "BMI", y = "sub-population", fill = "BMI") +
#   theme(legend.position="right") 
# 
# p2 = mydata %>% ggplot(aes(y = subpop, x = Yengo_Weighted_GRS, fill = stat(x))) + 
#   geom_density_ridges_gradient(scale = 2, rel_min_height = 0.0001) +
#   scale_fill_gradient2(low = pcol[1], mid = pcol[2], high = pcol[3], midpoint = 10.2) +
#   labs(x = "BMI-GRS", y = "sub-population", fill = "BMI-GRS") +
#   theme(legend.position="right") 
# 
# p3 = mydata %>% ggplot(aes(y = subpop, x = leeftijd, fill = stat(x))) + 
#   geom_density_ridges_gradient(scale = 2, rel_min_height = 0.0001) +
#   scale_fill_gradient2(low = pcol[1], mid = pcol[2], high = pcol[3], midpoint = 55) +
#   labs(x = "age", y = "sub-population", fill = "age") +
#   theme(legend.position="right") 
# 
# p4 = mydata %>% ggplot(aes(y = subpop, x = opleid, fill = stat(x))) + 
#   geom_density_ridges_gradient(scale = 2, rel_min_height = 0.0001) +
#   scale_fill_gradient2(low = pcol[1], mid = pcol[2], high = pcol[3], midpoint = 5) +
#   labs(x = "educational attainment", y = "sub-population", fill = paste0("educational\nattainment")) +
#   theme(legend.position="right") 
# 
# p5 = mydata %>% ggplot(aes(y = subpop, x = sp2meth, fill = stat(x))) + 
#   geom_density_ridges_gradient(scale = 2, rel_min_height = 0.0001) +
#   scale_fill_gradient2(low = pcol[1], mid = pcol[2], high = pcol[3], midpoint = 5) +
#   labs(x = "educational attainment", y = "sub-population", fill = paste0("educational\nattainment")) +
#   theme(legend.position="right") +
#   xlim(0, 25)
# 
# grid.arrange(p1, p2, p3, ncol = 1)
  
```


## Distribution of BMI by sex.


```{r, fig.width = 8, fig.height = 6}
pcol = RColorBrewer::brewer.pal(5, "Set1")


w = which( mydata$sex == "male")
a = shapiro.test( mydata$bmi[w] )
p1 = mydata[w,] %>% ggplot(aes(bmi)) + 
  geom_histogram(fill = pcol[1], color = "grey" ,  binwidth = 1) +
  geom_vline(xintercept = mean(mydata$bmi[w], na.rm = TRUE), color = "grey", size = 2 ) +
  labs(x = "BMI", title = "Male BMI Distribution", subtitle = paste0( "W-stat = ", signif(a$sta, d = 3), "; N = ", length(w) ))
##

w = which(mydata$sex == "female")
a = shapiro.test( mydata$bmi[w] )
p2 = mydata[w,] %>% ggplot(aes(bmi)) + 
  geom_histogram(fill = pcol[2], color = "grey", binwidth = 1 ) +
  geom_vline(xintercept = mean(mydata$bmi[w], na.rm = TRUE), color = "grey", size = 2 ) +
  labs(x = "BMI", title = "Female BMI Distribution", subtitle = paste0( "W-stat = ", signif(a$sta, d = 3), "; N = ", length(w) ))
##
##########################
## Density Ridges
##
##########################
pcol = RColorBrewer::brewer.pal(8, "Set1")[3:1]
pcol = wesanderson::wes_palette("Darjeeling1", 3, type = c("discrete") )
#pcol = wesanderson::wes_palette("BottleRocket2", 3, type = c("discrete") )

w = which( !is.na( mydata$subpop) )

ridge_plot = mydata[w, ] %>% ggplot( aes(x = bmi, y = as.factor(sex), fill = stat(x) ) ) + 
  # geom_density_ridges(alpha=0.8, scale = 2) +
  geom_density_ridges_gradient(scale = 2, rel_min_height = 0.0001) +
  #scale_fill_viridis_c( option = "city", begin = 0.3, end = 1) + 
  scale_fill_gradient2(low = pcol[1], mid = pcol[2], high = pcol[3], midpoint = 30) +
  labs( fill = "BMI",  x = "BMI", y = "Sex Sample", title = "Variation in BMI by sex",
    subtitle = "Male and Female sub-populations")


mat = matrix(c(1,2,3,3), nrow = 2, byrow = TRUE)
 
## Print figure to file
# w = which(par_data[,1] == "results_dir")
# f = paste0( par_data[w,2] , "figures/2b.BMI_distributions_by_sex.pdf")
# pdf(f, width = 8, height = 6)
# grid.arrange(p1, p2, ridge_plot, layout_matrix = mat  )
# dev.off()

grid.arrange(p1, p2, ridge_plot, layout_matrix = mat  )

```

## Structure of sampling date with the two sub-populations

```{r}

table(mydata$subpop, mydata$visit_date)

```

## Do the Principal Components correlate with the subsampling structure?

```{r}
v = c("PC1","PC2","PC3","PC4")
PC_test = t( sapply(v, function(x){
  form = as.formula( paste0( x, " ~ as.factor(subpop)" ) )
  fit = lm(form, data = mydata)
  out = c( summary(fit)$r.squared, summary(fit)$coefficients[2,4] )
  names(out) = c("R2","P")
  return(out)
}) )

PC_test

```

**PC3 and PC4 correlate with sub-population structure**

## PCA 3 and 4 by subpopulations

```{r, fig.width = 7, fig.height = 4 }
pcol = RColorBrewer::brewer.pal(4, "Set1")

L = which(mydata$subpop == "Leiderdorp")
O = which(mydata$subpop == "Leiden")
####
p1 = as_tibble(mydata[c(O, L), ]) %>% ggplot( aes(x = PC3, y = PC4) ) +
  geom_point(aes(color = as.factor(subpop), shape = as.factor(subpop)  ), size = 3) +
  scale_color_manual(values=pcol) +
  labs(shape = "population", color = "population", 
       x = paste0("PC3, Rsq with subpop = ", round(PC_test[3,1], d = 3), ", P = ", signif(PC_test[3,2], d = 3) ),
       y = paste0("PC4, Rsq with subpop = ", round(PC_test[4,1], d = 3), ", P = ", signif(PC_test[4,2], d = 3) ))


## Print figure to file
# w = which(par_data[,1] == "results_dir")
# f = paste0( par_data[w,2] , "figures/3.PC3_PC4_correlate_w_subPOPs.pdf")
# pdf(f, width = 7, height = 4)
# p1
# dev.off()

p1

```

## Distribution of W-stats among the metabolite traits

```{r metabolite_Wstats, fig.width = 8, fig.height = 5, warning=FALSE, error=FALSE, message=FALSE}
fast = traits[grep("_f", traits)]
post = traits[grep("_p", traits)]
respo = traits[grep("_r", traits  )]

######
Wstats = t( apply(mydata[, traits], 2, function(x){
  y = sample(x, 5000)
  sh = shapiro.test(y)
  out = c(sh$statistic, sh$p.value)
  return(out)
}) )
####
Wstats = as.data.frame(Wstats)
Wstats$metabolite = rownames(Wstats)
colnames(Wstats) = c("W","P","metabolite")
###
Wstats$dietarystate = "fasting"
Wstats$dietarystate[ Wstats$metabolite %in% post] = "postprandial"
Wstats$dietarystate[ Wstats$metabolite %in% respo] = "response"
###
p1 = Wstats %>% ggplot(aes( W )) + 
  geom_histogram(fill = "grey", color = "grey30") +
  geom_vline(xintercept = mean(Wstats$W, na.rm = TRUE), color = "grey20", size = 2 ) +
  labs(x = "W-statistic", title = paste0("Distribution of metabolite W-statistics") )

p2 = Wstats %>% ggplot(aes( W )) + 
  geom_histogram( aes(fill = dietarystate), color = "grey30") +
  labs(x = "W-statistic", title = paste0("Distribution of metabolite W-statistics by dietary state") ) +
  facet_wrap(~ dietarystate) + 
  scale_fill_manual(values = pcol[1:3])

############
# w = which(par_data[,1] == "results_dir")
# f = paste0( par_data[w,2] , "figures/4.Metabolite_Wstats.pdf")
# pdf(f, width = 8, height = 5)
# grid.arrange(p1, p2, nrow = 2)
# dev.off()
# 

grid.arrange(p1, p2, nrow = 2)

```



## Observational correlation between BMI and BMI-GRS by population sample

```{r bmigrs-bmi, echo = FALSE, fig.width = 5, fig.height = 2}

quick_stats = function(wdata, pweights=NA){
  if( is.na(pweights[1]) ){
    fit = lm(bmi ~ Yengo_Weighted_GRS, data = wdata)  
  } else {
    fit = lm(bmi ~ Yengo_Weighted_GRS, data = wdata, weights = pweights)  
  }
  n = length(residuals(fit))
  s = summary(fit)
  coef = s$coefficients[2,c(1,2,3,4)]
  r2 = s$r.squared
  out = c(n, r2, coef)
  names(out) = c("n","r2","beta","se","tval", "P")
  return(out)
}
##########################################
## Estimate the effect of BMI-GRS on BMI
##########################################
BMI_GRS_Estimates = c()
## NEO
x = quick_stats(wdata = mydata, pweights = NA)
BMI_GRS_Estimates = rbind(BMI_GRS_Estimates, "NEO" = x)
## MALES
w = which(mydata$sex == "male")
x = quick_stats(wdata = mydata[w, ], pweights = NA)
BMI_GRS_Estimates = rbind(BMI_GRS_Estimates, "male" = x)
## FEMALES
w = which(mydata$sex == "female")
x = quick_stats(wdata = mydata[w, ], pweights = NA)
BMI_GRS_Estimates = rbind(BMI_GRS_Estimates, "female" = x)
## Leiderdorp
w = grep("Leiderdorp", mydata$subpop)
x = quick_stats(wdata = mydata[w, ], pweights = NA)
BMI_GRS_Estimates = rbind(BMI_GRS_Estimates, "Leiderdorp" = x)
## Leiden
w = grep("Leiden", mydata$subpop)
x = quick_stats(wdata = mydata[w, ], pweights = NA)
BMI_GRS_Estimates = rbind(BMI_GRS_Estimates, "Leiden" = x)
## Weighted NEO
x = quick_stats(wdata = mydata, pweights = mydata$pweight)
BMI_GRS_Estimates = rbind(BMI_GRS_Estimates, "weighted_NEO" = x)

BMI_GRS_Estimates = as.data.frame(BMI_GRS_Estimates)

BMI_GRS_Estimates %>% kbl() %>%
  kable_classic(full_width = F, html_font = "Cambria")

```

## Is the GRS associated with sex, age, smoking, alcohol ?

```{r}
cat(paste0("associated with sex ?\n"))
fit = glm(sex ~ Yengo_Weighted_GRS, data = mydata, weights = mydata$pweights, family = binomial)
summary(fit)$coef[2,]

cat(paste0("\nassociated with age ?\n"))
fit = lm(age ~ Yengo_Weighted_GRS, data = mydata, weights = mydata$pweights)
summary(fit)$coef[2,]

cat(paste0("\nassociated with higher educational attainment (ever/never)?\n"))
fit = glm(edu_higher ~ Yengo_Weighted_GRS, data = mydata, weights = mydata$pweights, family = binomial)
summary(fit)$coef[2,]

cat(paste0("\nassociated with smoking (ever/never)?\n"))
fit = glm(smoking_ever_never ~ Yengo_Weighted_GRS, data = mydata, weights = mydata$pweights, family = binomial)
summary(fit)$coef[2,]

cat(paste0("\nassociated with smoking (pack years) ?\n"))
fit = lm(packyears ~ Yengo_Weighted_GRS, data = mydata, weights = mydata$pweights)
summary(fit)$coef[2,]

cat(paste0("\nassociated with alcohol ?\n"))
fit = lm(alcohol ~ Yengo_Weighted_GRS, data = mydata, weights = mydata$pweights)
summary(fit)$coef[2,]

##############

cat(paste0("\nassociated with visit data ?\n"))
fit = glm(visit_date ~ Yengo_Weighted_GRS, data = mydata, weights = mydata$pweights, family = binomial)
summary(fit)$coef[2,]

cat(paste0("\nassociated with subpop ?\n"))
fit = glm(subpop ~ Yengo_Weighted_GRS, data = mydata, weights = mydata$pweights, family = binomial)
summary(fit)$coef[2,]

cat(paste0("\nassociated with PC3 ?\n"))
fit = lm(PC3 ~ Yengo_Weighted_GRS, data = mydata, weights = mydata$pweights)
summary(fit)$coef[2,]

cat(paste0("\nassociated with PC4 ?\n"))
fit = lm(PC4 ~ Yengo_Weighted_GRS, data = mydata, weights = mydata$pweights)
summary(fit)$coef[2,]


cat(paste0("\nassociated with being on a diet (always|sometimes|no) ?\n"))
fit = lm(as.numeric(diet) ~ Yengo_Weighted_GRS, data = mydata, weights = mydata$pweights)
summary(fit)$coef[2,]

cat(paste0("\nassociated with being on a weightloss diet (yes|no) ?\n"))
fit = glm(diet_weigth_loss ~ Yengo_Weighted_GRS, data = mydata, weights = mydata$pweights, family = binomial)
summary(fit)$coef[2,]





```


## Observational correlation between BMI and BMI-GRS

```{r bmigrs-bmi, echo = FALSE, fig.width = 7.5, fig.height = 1.5}
w = grep("_GRS", colnames(mydata))
grs_vars = colnames(mydata)[w]
##
bmi_cor = t( sapply(grs_vars, function(i){
	a = cor.test(mydata$bmi, mydata[,i])
	out = c(a$estimate, a$estimate^2, a$p.value)
	return(out)
	}) ) 

colnames(bmi_cor) = c("r","r2","pval")

grid.arrange( tableGrob( bmi_cor ) )

```

### Observational correlation between BMI and BMI-GRS in Leiderdorp sub-sample

```{r bmigrs-bmi_Leiderdrop, echo = FALSE, fig.width = 7.5, fig.height = 1.5}
w = grep("_GRS", colnames(mydata))
grs_vars = colnames(mydata)[w]
##
ss = which(mydata$subpop == "Leiderdorp")
m = mean(mydata$bmi[ss])
##
bmi_cor = t( sapply(grs_vars, function(i){
	a = cor.test(mydata[ss, "bmi"], mydata[ss,i])
	out = c(a$estimate, a$estimate^2, a$p.value)
	return(out)
	}) ) 

colnames(bmi_cor) = c("r","r2","pval")

grid.arrange( tableGrob( bmi_cor ) )

```

### Observational correlation between BMI and BMI-GRS in Leiden sub-sample

```{r bmigrs-bmi_Obese, echo = FALSE, fig.width = 8, fig.height = 1.5}
w = grep("_GRS", colnames(mydata))
grs_vars = colnames(mydata)[w]
##
ss = which(mydata$subpop == "Leiden")
##
bmi_cor = t( sapply(grs_vars, function(i){
	a = cor.test(mydata[ss, "bmi"], mydata[ss,i])
	out = c(a$estimate, a$estimate^2, a$p.value)
	return(out)
	}) ) 

colnames(bmi_cor) = c("r","r2","pval")

grid.arrange( tableGrob( bmi_cor ))

```


## Observational correlation between rank normal transformed BMI and BMI-GRS

```{r rnt_bmigrs-bmi, echo = FALSE, fig.width = 8, fig.height = 1.5}
w = grep("_GRS", colnames(mydata))
grs_vars = colnames(mydata)[w]
##
rnt_bmi_cor = t( sapply(grs_vars, function(i){
	a = cor.test( rntransform( mydata$bmi ), mydata[,i])
	out = c(a$estimate, a$estimate^2, a$p.value)
	return(out)
	}) ) 
colnames(rnt_bmi_cor) = c("r","r2","pval")


grid.arrange( tableGrob( rnt_bmi_cor ) )

```

```{r bmigrs_to_bmi_plot, fig.width = 8, fig.height = 4, warning=FALSE, error=FALSE, message=FALSE}

## Plot
pcol = RColorBrewer::brewer.pal(9, "Set1")
#pcol = wesanderson::wes_palette(name ="FantasticFox1", n = 4)
##
p1 = mydata %>% ggplot(aes(x = bmi, y = Yengo_Weighted_GRS)) +
	geom_point(fill = pcol[2], size = 2.5, shape = 21) + 
	geom_smooth(formula = y~x, method = "loess", color = pcol[5], size = 1.5) +
	geom_smooth(formula = y~x, method = "lm", color = pcol[1], size = 1.5) +
	labs(title = "BMI, BMI-GRS correlation", x = "BMI", y = "Yengo weighted BMI-GRS", subtitle = paste0("r2 = ", signif(bmi_cor[4,2], d = 5) ) )


p2 = mydata %>% ggplot(aes(x = rntransform(bmi), y = Yengo_Weighted_GRS)) +
	geom_point(fill = pcol[3], size = 2.5, shape = 21) + 
	geom_smooth(formula = y~x, method = "loess", color = pcol[5], size = 1.5) +
	geom_smooth(formula = y~x, method = "lm", color = pcol[1], size = 1.5) +
	labs(title = paste0("rnt-BMI, BMI-GRS correlation"), x = "RNT(BMI)", y = "Yengo weighted BMI-GRS", subtitle = paste0("r2 = ", signif(rnt_bmi_cor[4,2], d = 5) ) )

### Print PDF to file
# w = which(par_data[,1] == "results_dir")
# f = paste0( par_data[w,2] , "figures/5.BMI_BMIGRS_cor.pdf")
# pdf(f, width = 8, height = 4)
# grid.arrange(p1, p2, nrow = 1)
# dev.off()

grid.arrange(p1, p2,  nrow = 1)


```

## Is BMI or BMI-GRS structured by visit/sampling day ?

```{r BMI_visitday, fig.width = 10, fig.height = 8, warning=FALSE, error=FALSE, message=FALSE}

## BMI
# fit = lm( rntransform(bmi) ~ visit_date, data = mydata)
fit = lm( bmi ~ visit_date, data = mydata)
varexp = signif( summary(fit)$r.squared *100, d = 4)
mean_bmi = mean(mydata$bmi, na.rm = TRUE)
## BMI-GRS
fit_grs = lm(Yengo_Weighted_GRS ~ visit_date, data = mydata)
varexp_grs = signif( summary(fit_grs)$r.squared *100, d = 4)
mean_grs = mean(mydata$Yengo_Weighted_GRS, na.rm = TRUE)
##
pcol = RColorBrewer::brewer.pal(5, "Set1")
###
plotA = mydata %>% ggplot( aes(x= visit_date, y=bmi)) + 
	geom_boxplot( fill = pcol[2]) + 
	theme(axis.text.x  = element_text(angle=90, vjust=0.5)) +
	geom_hline(yintercept = mean_bmi, color = pcol[1], size = 1.5) +
	labs(x = "Sampling Year-Month", y = "BMI", 
		title = "Is BMI structured by sampling date?",
		subtitle = paste0("\t",varexp, "% of variance in BMI explained by sampling date") )
###
plotB = mydata %>% ggplot( aes(x= visit_date, y= Yengo_Weighted_GRS )) + 
	geom_boxplot( fill = pcol[3]) + 
	theme(axis.text.x  = element_text(angle=90, vjust=0.5)) +
	geom_hline(yintercept = mean_grs, color = pcol[1], size = 1.5) +
	labs(x = "Sampling Year-Month", y = "BMI-GRS", 
		title = "Is BMI-GRS structured by sampling date?",
		subtitle = paste0("\t",varexp_grs, "% of variance in BMI-GRS explained by sampling date") )


###
# w = grep("results_dir", par_data[,1])
# f = paste0(par_data[w,2], "figures/6.BMI_GRS_samplingmonth_boxplot.pdf")
# pdf(f, width = 10, height = 10)
# grid.arrange(plotA, plotB, nrow = 2)
# dev.off()
#####

grid.arrange(plotA, plotB, nrow = 2)

```

### What is the influence of visit date and BMI on metabolite abundance

Here I am running three linear models on rank normal transformed metabolite abundances, with (1) visit date, (2) BMI, and (3) visit date and BMI as covariates. 

```{r visit_metabolite_bmi_cors, echo = FALSE}
## define my list of interested traits (metabolites)
# traits = c(fast, post, respo)
################################################
## estimate correlations and variance explained
################################################
visit_cors = t( 
	sapply(traits, function(met){
	## W-stat
	W = signif( shapiro.test(sample( mydata[,met], 5000 ) )$stat , d = 4 )
	names(W) = "W-stat"
	####################
	## VISIT UNIVARIATE
	####################
	form = as.formula( paste0( "rntransform(",met,")", " ~ visit_date") )
	fit =  lm( form, data = mydata )  
	varexp = signif( summary(fit)$r.squared , d = 4)
	names(varexp) = "uv_visit_varexp"
	a = anova(fit)
	pval = a[1,5]
	names(pval) = "uv_visit_pval"
	####################
	## BMI UNIVARIATE
	####################
	form = as.formula( paste0( "rntransform(", met ,")", " ~ bmi") )
	fitB =  lm( form, data = mydata )  
	varexp_B = signif( summary(fitB)$r.squared , d = 4)
	names(varexp_B) = "uv_bmi_varexp"
	a = anova(fitB)
	pval_B = a[1,5]
	names(pval_B) = "uv_bmi_pval"
	####################
	## VISIT BMI MULTIVARIATE
	####################
	form = as.formula( paste0( "rntransform(", met ,")", " ~ visit_date + bmi") )
	fitM =  lm( form, data = mydata )  
	varexp_M = signif( summary(fitM)$r.squared , d = 4)
	a = anova(fitM)
	varexp_M = a[1:2,2]/sum(a[,2])
	names(varexp_M) = paste0("mv_bmi_varexp_", rownames(a)[1:2])
	pval_M = a[1:2,5]
	names(pval_M) = paste0("mv_bmi_pval_", rownames(a)[1:2])
	####################
	## BMI-GRS UNIVARIATE
	####################
	form = as.formula( paste0( "rntransform(", met ,")", " ~ Yengo_Weighted_GRS") )
	fitBGRS =  lm( form, data = mydata )  
	varexp_BGRS = signif( summary(fitBGRS)$r.squared , d = 4)
	names(varexp_BGRS) = "uv_bmiGRS_varexp"
	a = anova(fitBGRS)
	pval_BGRS = a[1,5]
	names(pval_BGRS) = "uv_bmiGRS_pval"
	####################
	## VISIT BMI-GRS MULTIVARIATE
	####################
	form = as.formula( paste0( "rntransform(", met ,")", " ~ visit_date + Yengo_Weighted_GRS") )
	fitMGRS =  lm( form, data = mydata )  
	varexp_MGRS = signif( summary(fitMGRS)$r.squared , d = 4)
	a = anova(fitMGRS)
	varexp_MGRS = a[1:2,2]/sum(a[,2])
	names(varexp_MGRS) = paste0("mv_bmiGRS_varexp_", rownames(a)[1:2])
	pval_MGRS = a[1:2,5]
	names(pval_MGRS) = paste0("mv_bmiGRS_pval_", rownames(a)[1:2])
	##
	out = c(W, varexp, pval, varexp_B, pval_B, varexp_M, pval_M, varexp_BGRS, pval_BGRS, varexp_MGRS, pval_MGRS )
	##
	return(out)
	}) 
)


visit_cors = as.data.frame(visit_cors)
visit_cors = as_tibble(visit_cors)
visit_cors$metabolite = traits

####
visit_cors$dietary_state = "response"
##
w = grep( "_p", visit_cors$metabolite)
visit_cors$dietary_state[w] = "postprandial"
#####
w = grep( "_f", visit_cors$metabolite)
visit_cors$dietary_state[w] = "fasting"


```

Plot of the variance explained on each metabolite (dots) by, bmi, visit date, and BMI corrected for visit date.

```{r met_BMI_cors, fig.width = 12, fig.height = 5, warning=FALSE, error=FALSE, message=FALSE}
pcol = RColorBrewer::brewer.pal(8, "Set1")
plotA = visit_cors %>% ggplot( aes( x = uv_visit_varexp, y = uv_bmi_varexp)  ) +
	geom_point( aes(size = -log10(uv_bmi_pval), 
	                color = dietary_state ) ,
	             alpha = 0.7 ) +
	labs(x = "Variance explained by year-month of visit",
		y = "Variance explained by BMI" , title = "Variance explained by visit date and BMI on metabolite abundance",
		subtitle = "- Univariate models -",
		size = "-log10(Univariate BMI)",
		color = "Dietary State" ) +
	geom_smooth( formula = 'y ~ x' , method = "lm", color = pcol[1]) +
  theme(legend.position="right") +
  #scale_colour_gradient2(low = "grey30", mid = pcol[2], high = pcol[3] )
  scale_fill_manual(values = pcol[1:3])

plotB = visit_cors %>% ggplot( aes( x = uv_bmi_varexp, y = mv_bmi_varexp_bmi)  ) +
	geom_point( aes(size = -log10(mv_bmi_pval_bmi), 
	                color = dietary_state ) , 
	            alpha = 0.7) +
	labs(x = "Variance explained by BMI",
		y = "Variance explained by BMI, corrected by visit date", title = "Variance explained by BMI on metabolite abundance",
		subtitle = "- Comparision of Univariate & Multivariate estimates -" , 
		size = "-log10(Multivariate BMI)",
		color = "Dietary State" ) +
	geom_smooth( formula = 'y ~ x' , method = "lm", color = pcol[1]) +
	geom_abline( slope = 1, intercept = 0, color = "black", size = 2 ) +
  theme(legend.position="right") +
  # scale_colour_gradient2(low = "grey30", mid = pcol[2], high = pcol[3] )
  scale_fill_manual(values = pcol[1:3])

####
# w = grep("results_dir", par_data[,1])
# f = paste0(par_data[w,2], "figures/7.BMI_VisitDate_VarExp.pdf")
# pdf(f, width = 14, height = 5)
# grid.arrange(plotA, plotB, nrow = 1)
# dev.off()
#####

grid.arrange(plotA, plotB, nrow = 1)


```

### Average influence on fasting, postprandial, and response metabolite abundance

```{r dietary_state_varexp, fig.width = 10, fig.height = 5, echo = FALSE, warning=FALSE, message=FALSE, error=FALSE}
##### Visit Date
plotA = visit_cors %>% ggplot(aes(y = uv_visit_varexp, x = dietary_state )) +
  geom_boxplot( aes(fill = dietary_state)) +
  labs(x = "Dietary State", y = "Variance Explained by Visit Date", 
       title = "Univ-Variance Explained in Metabolite by Visit Date") +
  theme(legend.position="bottom")

## BMI
plotB = visit_cors %>% ggplot(aes(y = uv_bmi_varexp, x = dietary_state )) +
  geom_boxplot( aes(fill = dietary_state)) +
  labs(x = "Dietary State", y = "Variance Explained by BMI", 
       title = "Univ-Variance Explained in Metabolite by BMI") +
  theme(legend.position="bottom")


####
# w = grep("results_dir", par_data[,1])
# f = paste0(par_data[w,2], "figures/8.VarExp_acrossMetabolites_byDate_byBMI.pdf")
# pdf(f, width = 10, height = 5)
# grid.arrange(plotA, plotB, nrow = 1)
# dev.off()
#####

grid.arrange(plotA, plotB, nrow = 1)

```

### Top 100 metabolites associated with visit date

```{r visit_date_metabolite,echo = FALSE}
o = order(visit_cors$uv_visit_pval, decreasing = FALSE)
ds = sapply( visit_cors$metabolite[o[1:100]], function(x){ strsplit( x , split = "_")[[1]][2] })
perc_response = (length( grep( "r", ds ) ) / 100) *100
###
visit_cors$metabolite[o[1:100]]

```

#### what % are response traits?
`r perc_response`% of the top 100 metabolites are response traits. 

### Visit dates impact on top 2 metabolites 

```{r XSVLDLCEp_1, echo = FALSE, fig.width = 10, fig.height = 6, message=FALSE, warning=FALSE, error=FALSE}
pcol = RColorBrewer::brewer.pal(8, "Set1")

fit = lm( rntransform( xsvldlcepct_f ) ~ visit_date, data = mydata)
a = anova(fit)
varexp = (a[1,2]/sum(a[,2]))*100
####
plotA = mydata %>% ggplot( aes(x= visit_date, y= rntransform( xsvldlcepct_f ) )) + 
	geom_boxplot( fill = pcol[5]) + 
	theme(axis.text.x  = element_text(angle=90, vjust=0.5)) +
	geom_hline(yintercept = mean( rntransform(mydata$xsvldlcepct_f) , na.rm = TRUE) , color = pcol[1], size = 1.5) +
	labs(x = "Sampling Year-Month", y = "Fasting XS-VLDL-CE %", 
		title = "Fasting XS-VLDL-CE % structured by sampling date",
		subtitle = paste0("\t",varexp, "% of variance in XS-VLDL-CEp explained by sampling date") )

#################

fit = lm( rntransform( xsvldlc_r ) ~ visit_date, data = mydata)
a = anova(fit)
varexp = (a[1,2]/sum(a[,2]))*100
####
plotB = mydata %>% ggplot( aes(x= visit_date, y= rntransform( xsvldlc_r ) )) + 
	geom_boxplot( fill = pcol[4]) + 
	theme(axis.text.x  = element_text(angle=90, vjust=0.5)) +
	geom_hline(yintercept = mean( rntransform(mydata$xsvldlc_r) , na.rm = TRUE) , color = pcol[1], size = 1.5) +
	labs(x = "Sampling Year-Month", y = "XS-VLDL-C response", 
		title = "XS-VLDL-C response structured by sampling date",
		subtitle = paste0("\t",varexp, "% of variance in XS-VLDL-CEp explained by sampling date") )

####
# w = grep("results_dir", par_data[,1])
# f = paste0(par_data[w,2], "figures/9.VisitDate_top2Metabolite_boxplot.pdf")
# pdf(f, width = 12, height = 10)
# grid.arrange(plotA, plotB, nrow = 2)
# dev.off()
#####

grid.arrange(plotA, plotB, nrow = 2)


```



### Top 100 metabolites associated with BMI

```{r bmi_metabolites, echo = FALSE}
o_bmi = order(visit_cors$uv_bmi_pval, decreasing = FALSE)
visit_cors$metabolite[o_bmi[1:100]]
```

### How many of the top 100 metabolites, for visit date and BMI overlap?
 - overlap count = `r sum( visit_cors$metabolite[o_bmi[1:100]] %in% visit_cors$metabolite[o[1:100]] )`


# Metabolite Clustering by visit date profile

Procedure: 
  1. exclude those metabolites not influenced by sample date
  2. rank transform each metabolite
  2. estimate the mean for each metabolite-X-sample date
  3. run kmeans

```{r}
## Identify Metabolites to retain for the analysis
BF = 0.05/43
w = which(visit_cors$uv_visit_pval < BF)
# w = which( visit_cors$uv_bmi_varexp > 0.03)
keep = traits[w]
vd = unique(mydata$visit_date)
vd = sort(vd)
## Estimate the mean estimates for each metabolite-X-SampleDate
met_visitdate_means = t( apply( mydata[,keep], 2, function(x){
  y = rntransform(x)
  m = sapply(vd, function(i){
    w = which(mydata$visit_date == i)
    mean( y[w], na.rm = TRUE)
  })
  return(m)
}) )

## Turn NAs into 0
met_visitdate_means[is.na(met_visitdate_means)] = 0
colnames(met_visitdate_means) = vd
```

```{r}
## Run Kmeans
set.seed(22052021)
k = lapply(2:20, function(k){
  out = kmeans( met_visitdate_means , centers = k, iter.max = 100, nstart = 25)
  return(out)
} )


```

## Estimate silhouette score and between cluster sums of squares

```{r}
###################
## Silhouette
###################
dmat = dist( met_visitdate_means )
##
silhouette_score = unlist( lapply(k, function(x){
  ss = silhouette(x$cluster, dmat)
  score = mean(ss[, 3])
}) )
###################
## BETWEEN Sums of Squares
###################
BSS = unlist( lapply(k, function(x){ x$betweenss}) )

```

### plot silhouette score and BSS for each k

```{r,echo = FALSE, fig.width = 8, fig.height = 5}
par(mfrow = c(1,2))
## silhouette
plot(2:20, silhouette_score, pch = 21, bg = "royalblue" ,  cex = 1.5, type='b',
     xlab='Number of clusters', ylab='Average Silhouette Scores', frame=FALSE,
     main = "Silhouette")
## Between SS
plot(2:20, BSS, pch = 21, bg = "dodgerblue", cex = 1.5, type = "b",
     xlab = "k", ylab = "between cluster sum of squares", frame=FALSE,
     main = "Between Sum of Squares")


```

## Cluster Profiles: 

### estimate the mean and 95% CI rank value across metabolites in a cluster for each sample date

```{r, fig.width = 8, fig.height = 4}
## Define my clusters
myK = k[[9]]$cluster
## Size of each cluster
n = table(myK)
## estimate the mean and 95% CI of each cluster for each sample year month
profiles = lapply(1:length(unique(myK)), function(j){
  w = which(myK == j)
  date_estimates = apply( met_visitdate_means[w, ], 2, function(x){
    m = mean(x)
    s = quantile(x, probs = c(0.025,0.975))
    out = c(m, s)
    return(out)
  })
  return(date_estimates)
})

# plot(1:49, profiles[[4]][1,], pch = 21, bg = "royalblue", cex = 2 )

mytable = c()
for(i in 1:length(profiles) ){
  x = t(profiles[[i]])
  colnames(x) = c("mean","lowerCI","upperCI")
  x = as.data.frame(x)
  x$cluster = paste0("cluster_",i, "_n=", n[i])
  x$date = rownames(x)
  mytable = rbind(mytable, x)
}
##
mytable$cluster = factor(mytable$cluster, levels = unique(mytable$cluster) )
mytable$date = factor(mytable$date, levels = sort( colnames(met_visitdate_means) ) )

```

### Plot cluster profiles across sample dates

```{r, fig.width = 12, fig.height = 6}
mytable %>% ggplot(aes( x = date, y = mean )) +
  geom_point(aes(color = as.factor(cluster) )) +
  geom_errorbar( aes(ymin= lowerCI , ymax= upperCI ), 
                 width=.2,
                 position=position_dodge(0.05), color = "grey50" ) +
  facet_wrap(. ~ as.factor(cluster) ) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1, size = 6)) +
  labs(color = "Kmeans Cluster", title = "Metabolites clustered by Sample Date Profile",
       y = "metabolite cluster means and 95% CI", x = "sample date (year-month)")


```

### Cluster enrichment

```{r}
#names( myK[myK == 9] )
clus_data = data.frame(metabolite = names(myK), cluster = myK, dietarystate = "response", met = names(myK))
##
w = grep("_f", clus_data$metabolite)
clus_data$dietarystate[w] = "fasting"
w = grep("_p", clus_data$metabolite)
clus_data$dietarystate[w] = "postprandial"
##
clus_data$met = gsub("_f","",clus_data$met)
clus_data$met = gsub("_p","",clus_data$met)
clus_data$met = gsub("_r","",clus_data$met)


```



```{r }
CE = lapply(1:10, function(i){
    w = which(clus_data$cluster == i)
    hgtest( allfeatures = clus_data$dietarystate, selectedfeatures = clus_data$dietarystate[w] )
})

ClusterEnrichment = c()
for(i in 1:length(CE) ){
    z = CE[[i]]
    z = cbind(paste0("cluster_", i), rownames(z),  z )
    colnames(z)[1:2] = c("cluster","class")
    ClusterEnrichment = rbind(ClusterEnrichment, z)
}
  
```

```{r, fig.width = 10, fig.height = 7, fig.align='center'}
grid.arrange(tableGrob(ClusterEnrichment[, c(1,2,3,5,6)]))
```
