==> 10_obs_results_summary.Rmd <==
---
title: "NEO BMI-Metabolite Exploring Obs Point Estimate Results"
author: "David Hughes"
date: "13/12/2022"
output: 
  pdf_document:
    number_sections: true
---


```{r setup, include=FALSE}
knitr::opts_chunk$set( echo = TRUE, error = FALSE, warning = FALSE, message = FALSE)

source("LOAD.R")

```

## Load point estimates

```{r}
w = which(par_data[,1] == "results_dir")
resultsdir = par_data[w,2]

f = paste0(resultsdir, "obs_tsls_estimates_Dec12_2022.Rdata")
load( f)

## long format data
f = paste0(resultsdir, "obs_tsls_estimates_long_Dec12_2022.Rdata")
load( file = f)

```

## load sensitivity data

```{r}
## sensitivity results
w = which(par_data[,1] == "results_dir")
resultsdir = par_data[w,2]
f = paste0(resultsdir, "tables/Dec12_2022_wNEO_SENSITIVITY_estimates.txt")
wNEO_sensitivity = read.table(f, header = TRUE, sep = "\t", as.is = TRUE)

## long format of sensitivity 
wNEO_sensitivity_long = obsmr_2_longformat(data = wNEO_sensitivity, 
                               pop_id = "wNEO_sa", 
                               alpha = 0.05/43, 
                               sandwich_estimates = FALSE)
```

## add sensitivity data to list

```{r}
obs_tsls_est[["wNEO_sa"]] = wNEO_sensitivity

obs_tsls_est_long[["wNEO_sa"]] = wNEO_sensitivity_long
```


```{r}
names(obs_tsls_est)
```


## add annotation to the results data frames

```{r}
for(i in 1:length(obs_tsls_est) ){
  obs_tsls_est[[i]] = cbind(obs_tsls_est[[i]], study_data$feature_anno[, c("dietary_state","raw.label","class","subclass", "derived_features")] )
}
```

# Observational Results

## weighted NEO results

```{r}
cat(paste0("weighted NEO nominally significant count\n"))
sum( obs_tsls_est$wNEO$Obs_P < 0.05 )
cat(paste0("weighted NEO BF significant count\n"))
sum( obs_tsls_est$wNEO$Obs_P < 0.05/43 )
```

## Parition of observational assocaiton by dietary state

```{r}
cat(paste0("weighted NEO nominally significant count\n"))
w = which( obs_tsls_est$wNEO$Obs_P < 0.05 )
table(study_data$feature_anno$dietary_state[w] )
# obs_tsls_est$wNEO %>% filter(Obs_P < 0.05) %>% summarize(state = unique(dietary_state), count = table(dietary_state))

cat(paste0("\n\nweighted NEO BF significant count\n"))
w = which( obs_tsls_est$wNEO$Obs_P < 0.05/43 )
table(study_data$feature_anno$dietary_state[w] )

```

```{r}
f =  which( obs_tsls_est$wNEO$Obs_P < 0.05/43 & obs_tsls_est$wNEO$dietary_state == "fasting")
f = obs_tsls_est$wNEO$outcome[f]
f = gsub("_f","",f)
p =  which( obs_tsls_est$wNEO$Obs_P < 0.05/43 & obs_tsls_est$wNEO$dietary_state == "postprandial")
p = obs_tsls_est$wNEO$outcome[p]
p = gsub("_p","",p)
```


## Mean, min, and max variance explained in associated cross-sectional analysis 

```{r}
# cat(paste0("mean, min, and max eta-square of cross-section associations\n"))

obs_tsls_est$wNEO %>% 
  filter(Obs_P < 0.05/43) %>% 
  summarise(mean = mean(Obs_devexp_by_exposure),
            min = min(Obs_devexp_by_exposure) ,
            max = max(Obs_devexp_by_exposure) )

```


## Mean, min, and max variance explained  BY DIETARY STATE in associated cross-sectional analysis 

```{r}
obs_tsls_est$wNEO %>% 
  filter(Obs_P < 0.05/43) %>% 
  group_by(dietary_state) %>%
  summarise(mean = mean(Obs_devexp_by_exposure),
            min = min(Obs_devexp_by_exposure) ,
            max = max(Obs_devexp_by_exposure) )
```


## weighted NEO results in females

```{r}
cat(paste0("weighted NEO nominally significant count\n"))
sum( obs_tsls_est$wNEO_female$Obs_P < 0.05 )
cat(paste0("weighted NEO BF significant count\n"))
sum( obs_tsls_est$wNEO_female$Obs_P < 0.05/43 )
```

## Parition of observational assocaiton by dietary state

```{r}
cat(paste0("weighted NEO nominally significant count\n"))
w = which( obs_tsls_est$wNEO_female$Obs_P < 0.05 )
table(study_data$feature_anno$dietary_state[w] )
cat(paste0("weighted NEO BF significant count\n"))
w = which( obs_tsls_est$wNEO_female$Obs_P < 0.05/43 )
table(study_data$feature_anno$dietary_state[w] )

```

## FEMALES: Mean, min, and max variance explained in associated cross-sectional analysis 

```{r}
obs_tsls_est$wNEO_female %>% 
  filter(Obs_P < 0.05/43) %>% 
  #group_by(dietary_state) %>%
  summarise(mean = mean(Obs_devexp_by_exposure),
            min = min(Obs_devexp_by_exposure) ,
            max = max(Obs_devexp_by_exposure) )

```


## FEMALES: Mean, min, and max variance explained BY DIETARY STATE in associated cross-sectional analysis 

```{r}
obs_tsls_est$wNEO_female %>% 
  filter(Obs_P < 0.05/43) %>% 
  group_by(dietary_state) %>%
  summarise(mean = mean(Obs_devexp_by_exposure),
            min = min(Obs_devexp_by_exposure) ,
            max = max(Obs_devexp_by_exposure) )
```


## weighted NEO results in males

```{r}
cat(paste0("weighted NEO nominally significant count\n"))
sum( obs_tsls_est$wNEO_male$Obs_P < 0.05 )
cat(paste0("weighted NEO BF significant count\n"))
sum( obs_tsls_est$wNEO_male$Obs_P < 0.05/43 )
```

## Parition of observational assocaiton by dietary state

```{r}
cat(paste0("weighted NEO nominally significant count\n"))
w = which( obs_tsls_est$wNEO_male$Obs_P < 0.05 )
table(study_data$feature_anno$dietary_state[w] )
cat(paste0("weighted NEO BF significant count\n"))
w = which( obs_tsls_est$wNEO_male$Obs_P < 0.05/43 )
table(study_data$feature_anno$dietary_state[w] )

```

## MALES: Mean, min, and max variance explained  in associated cross-sectional analysis 

```{r}
obs_tsls_est$wNEO_female %>% 
  filter(Obs_P < 0.05/43) %>% 
  #group_by(dietary_state) %>%
  summarise(mean = mean(Obs_devexp_by_exposure),
            min = min(Obs_devexp_by_exposure) ,
            max = max(Obs_devexp_by_exposure) )
```


## MALES: Mean, min, and max variance explained BY DIETARY STATE in associated cross-sectional analysis 

```{r}
obs_tsls_est$wNEO_male %>% 
  filter(Obs_P < 0.05/43) %>% 
  group_by(dietary_state) %>%
  summarise(mean = mean(Obs_devexp_by_exposure),
            min = min(Obs_devexp_by_exposure) ,
            max = max(Obs_devexp_by_exposure) )
```

## male vs females variance explained

```{r}
cat(paste0("mean absolute cross-section beta in females\t"))
mean( abs(obs_tsls_est$wNEO_female$Obs_beta) )
cat(paste0("mean absolute cross-section beta in males\t"))
mean( abs(obs_tsls_est$wNEO_male$Obs_beta) )

cat(paste0("\nmean cross-section eta-squared in females\t"))
mean(obs_tsls_est$wNEO_female$Obs_devexp_by_exposure)
cat(paste0("mean cross-section eta-squared in males\t"))
mean(obs_tsls_est$wNEO_male$Obs_devexp_by_exposure)

```

## plot of sex specific estimates

```{r}
sexdata = data.frame(outcome = obs_tsls_est$wNEO_male$outcome,
                     male = obs_tsls_est$wNEO_male$Obs_beta,
                     maleP = obs_tsls_est$wNEO_male$Obs_P,
                     female = obs_tsls_est$wNEO_female$Obs_beta,
                     femaleP = obs_tsls_est$wNEO_female$Obs_P, 
                     state = obs_tsls_est$wNEO_female$dietary_state)
###
sexdata$rawlabel = study_data$feature_anno$raw.label
sexdata$class = study_data$feature_anno$class
###
sexdata$sig = sapply(1:nrow(sexdata), function(i){
  if(sexdata$maleP[i] < 0.05/43 & sexdata$femaleP[i] < 0.05/43){
    out = "both"
  } else {
    if(sexdata$maleP[i] < 0.05/43 & sexdata$femaleP[i] > 0.05/43){
      out = "male"
    } else {
      if(sexdata$maleP[i] > 0.05/43 & sexdata$femaleP[i] < 0.05/43){
        out = "female"
      } else {
        out = "neither"
      }
    }
  }
})

## correlation
a = cor.test(sexdata$female, sexdata$male)
## colors
pcol = RColorBrewer::brewer.pal(4, "Set1")

## plot
library(ggrepel)
##
(male_female_scatter_plot = sexdata %>% ggplot(aes(x = female, y = male)) +
  geom_hline(yintercept = 0, color = "grey", linetype = "dashed") +
  geom_vline(xintercept = 0,color = "grey", linetype = "dashed") +
  geom_point(aes(color = sig, shape = state), size = 3, alpha = 0.6) +
  scale_color_brewer(palette = "Set1") +
  scale_shape_manual(values = c(19,15,17) ) +
  geom_abline(slope = 1, intercept = 0, linetype = "dotted") +
  geom_smooth(method = "lm", formula = y~x, color = "black") +
  geom_text_repel( data = sexdata %>% filter(sig == "female" & female < 0 & male > 0), 
                  aes(label=outcome), 
                  min.segment.length = 0,
                  box.padding = 0.25, 
                  color = pcol[2], nudge_x = -0.03, nudge_y = 0.01 ) +
  geom_text_repel( data = sexdata %>% filter(sig == "male" & female < 0 & male > 0), 
                  aes(label=outcome), 
                  min.segment.length = 0,
                  box.padding = 0.25, 
                  color = pcol[3], nudge_x = -0.01, nudge_y = 0.03 ) +
  geom_text_repel( data = sexdata %>% filter(sig == "male" & female > 0 & male < 0), 
                  aes(label=outcome), 
                  min.segment.length = 0,
                  box.padding = 0.25, 
                  color = pcol[3], nudge_x = 0.01, nudge_y = -0.03 ) +
  labs(x = "female cross-sectional effect estimates", 
       y = "male cross-sectional effect estimates",
       color = "association") +
  annotate("text", x = 0.0625, y = -0.095, label = paste0("Pearson's r = ", round(a$estimate, d = 3) ) ) +
  theme_bw()
)

```


```{r}
f = paste0(resultsdir, "figures/male_female_cross_sectional_scatter.pdf")
pdf(file = f, width = 8, height = 6)
male_female_scatter_plot
dev.off()
```


```{r}
table(sexdata$sig)
```

## Associations in both but ordered by FEMALE effect size

```{r}
sexdata %>% filter( sig == "both" ) %>%
  arrange(desc(abs(female))) %>%
  kable() %>% kableExtra::kable_classic()
```



## Associations in both but ordered by MALE effect size

```{r}
sexdata %>% filter( sig == "both" ) %>%
  arrange(desc(abs(male))) %>%
  kable() %>% kableExtra::kable_classic()
```


## Female association with a different sign in males

```{r}
sexdata %>% filter(femaleP < 0.05/43 & female < 0 & male > 0) %>%
  kable() %>% kableExtra::kable_classic()
```


## Male association with a different sign in females

```{r}
sexdata %>% filter(maleP < 0.05/43 & female < 0 & male > 0) %>%
  kable() %>% kableExtra::kable_classic()
```

## Male association with a different sign in females

```{r}
sexdata %>% filter(maleP < 0.05/43 & female > 0 & male < 0) %>%
  kable() %>% kableExtra::kable_classic()
```


## FASTING DATA SUMMARY


## order by p-value

```{r}
temp = obs_tsls_est$wNEO %>% filter(dietary_state == "fasting") %>%
  dplyr::select(outcome, Obs_beta, Obs_se, Obs_P,Obs_devexp_by_exposure, raw.label, class, subclass) %>%
  mutate(zscore = Obs_beta/Obs_se) %>%
  arrange(desc(-log10(Obs_P)))

temp %>% kable() %>% kableExtra::kable_classic()
```

## effect estiamtes without a RNT

```{r}
tr = c("gp_f", "ile_f", "leu_f", "val_f")
##
x = t(sapply( tr, function(x){
  obsmr( wdata = study_data$working_data,
            outcome = x,
            exposure = "bmi",
            instrument = "Yengo_Weighted_GRS",
            covariates = c("visit_date","subpop","age","sex"),
            weights = "pweight",
            rnt_outcome = FALSE,
            outlier_method = "iqr",
            outlier_cutoff = 10,
            messages = TRUE)
}))
x = as.data.frame(x)
x %>% kable() %>% kableExtra::kable_classic()
```

## effect estimates by class and dietary state ; arranged by dietary state

```{r}
obs_tsls_est$wNEO %>% 
  group_by(dietary_state, class) %>% summarise(n = length(Obs_beta), 
                                abs_mean = mean(abs(Obs_beta)), 
                                min = min(Obs_beta), 
                                max = max(Obs_beta) ) %>%
  arrange(dietary_state, desc(abs_mean)) %>% 
  # arrange( desc(abs_mean)) %>% 
  kable() %>% kableExtra::kable_classic()
```

## effect estimates by class and dietary state ; arranged by absolute mean

```{r}
obs_tsls_est$wNEO %>% filter(Obs_P < 0.05/43) %>%
  group_by(dietary_state, class) %>% summarise(n = length(Obs_beta), 
                                abs_mean = mean(abs(Obs_beta)), 
                                min = min(Obs_beta), 
                                max = max(Obs_beta) ) %>%
  # arrange(dietary_state, desc(abs_mean)) %>% 
  arrange( desc(abs_mean)) %>% 
  kable() %>% kableExtra::kable_classic()
```


```{r, fig.width = 10, fig.height = 8, error=FALSE, warning=FALSE}

## Function to extract needed data
wdata = lipoprotein_wdata(wd = obs_tsls_est$wNEO, define_Dstate = "fasting")

## plot order of subclasses
wdata$trait = factor(wdata$trait , levels = c("concentration", "total_lipids", 
                                              "phospholipids", "free_cholesterol", 
                                              "total_cholesterol", "cholesterol_esters", 
                                              "triglycerides",
                                              "phospholipids/total_lipids", "free_cholesterol/total_lipids", 
                                              "total_cholesterol/total_lipids" , 
                                              "cholesterol_esters/total_lipids", 
                                              "triglycerides/total_lipids")[12:1] )

###
x = RColorBrewer::brewer.pal(8,"Set1")
pcol = c( rep(x[3],6) , x[4] , rep(x[1], 3), rep(x[2],4) ) 
###
plotA = wdata %>% ggplot(aes(x = subclass, y = trait, fill = Obs_beta)) +
  geom_tile() + 
  geom_text(data = subset(wdata, Obs_P <= 0.05/43), aes(label =  round(Obs_beta, d = 4) ), size = 2.4 ) +
  # geom_text(data = subset(wdata, MR_P <= 0.05/43), aes(label =  paste0(round(MR_beta, d = 4), "*" ) ), size = 2.4 ) +
  scale_fill_gradient2(low = "red", mid = "white", high = "blue") +
  theme(axis.text.x = element_text(angle = 35, vjust = 1, hjust=1, colour = pcol)) +
  labs(x = "lipoprotein subclass", y = "trait", fill = "beta") +
  facet_grid(value ~ ., scales = "free_y", space = "free_y")

plotB = wdata  %>% 
  filter(!grepl("/",trait)) %>% 
  ggplot(aes(x = as.numeric(subclass), y = Obs_beta, color = trait)) +
  geom_hline(yintercept = 0, size = 1, linetype = "dashed", color = "grey30") +
  geom_point(size = 4, alpha = 0.6) +
  geom_line( aes(color = trait) ) +
  labs(x = "lipoprotein subclass", y = "beta", color = "trait") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 40, vjust = 1, hjust=1, color = pcol)) + 
  scale_x_continuous(name = "lipoprotein subclass", breaks = 1:14, labels = levels(wdata$subclass) )

ggpubr::ggarrange(plotA, plotB, ncol = 1)  

```


## save fasting lipoproteins beta profile

```{r}
f = paste0(resultsdir, "figures/wNEO_fasting_lipo_Obs_profile.pdf")
pdf(file = f, width = 10, height = 8)
ggpubr::ggarrange(plotA, plotB, ncol = 1)  
dev.off()
```



## box plot of point estimates by class and dietary state

```{r, fig.width = 15, fig.height = 7}
p =  obs_tsls_est$wNEO %>%  ggplot( aes(x=class, y = Obs_beta) ) + 
  geom_boxplot( aes(fill = class) ) +
  geom_jitter( width = 0.15 ) +
  geom_hline(yintercept = 0) +
  facet_wrap(~dietary_state, nrow = 3, scales = "free") +
  theme_bw() +
  theme(axis.ticks.x = element_blank(), axis.text.x = element_blank() ) 
p
```


```{r}
f = paste0(resultsdir, "figures/Obs_beta_byclass_v3.pdf")
pdf(file = f, width = 12, height = 5)
p
dev.off()
```


### Comparison with Wurtz et al

```{r}
## define data dir
w = which(par_data[, 1] == "data_dir")
data_dir = par_data[w, 2]
## define file
f = paste0(data_dir, "Wurtz/Wurtz_Edited_Data.txt")
wurtz = read.table(f, header = TRUE, sep = "\t")
colnames(wurtz) = paste0("wurtz_", colnames(wurtz))
colnames(wurtz)[1] = "wurtz_longname"
## match to my fasting data
ftemp = obs_tsls_est$wNEO %>% filter(dietary_state == "fasting")
ftemp$metids = metids

m = match(ftemp$metids, wurtz$wurtz_ID)
tempd = cbind(ftemp[, c(1,2,17,18,20,67,68,70)], wurtz[m,])
rm(ftemp)
neo_wurtz = na.omit(tempd)

m = match(neo_wurtz$outcome, study_data$feature_anno$ids)
neo_wurtz = cbind(neo_wurtz, study_data$feature_anno[m, c("raw.label", "label", "label.no.units", "class", "subclass")])

dim(neo_wurtz)
```

```{r}
x = neo_wurtz$wurtz_obs_beta / neo_wurtz$wurtz_conversion_factor
a = cor.test(neo_wurtz$Obs_beta, x, method = "pearson")
# cor.test(neo_wurtz$Obs_beta, x, method = "spearman")

pdata = cbind( neo_wurtz[, c("outcome","Obs_beta")] , 
               Obs_LCI = neo_wurtz$Obs_beta - (1.96 * neo_wurtz$Obs_se),
               Obs_UCI = neo_wurtz$Obs_beta + (1.96 * neo_wurtz$Obs_se),
               Obs_P = neo_wurtz$Obs_P,
               wurtz_obs_beta = neo_wurtz$wurtz_obs_beta / neo_wurtz$wurtz_conversion_factor, 
               wurtz_obs_LCI = neo_wurtz$wurtz_obs_LCI / neo_wurtz$wurtz_conversion_factor, 
               wurtz_obs_UCI = neo_wurtz$wurtz_obs_UCI / neo_wurtz$wurtz_conversion_factor, 
               wurtz_obs_P = neo_wurtz$wurtz_obs_P
               )

#pdata$sig_both = ifelse(pdata$Obs_P < 0.05/43 & pdata$wurtz_obs_P < 0.05/100, "sig_both", "not")
pdata$sig = sapply(1:nrow(pdata), function(i){
  if(pdata$Obs_P[i] < 0.05/43 & pdata$wurtz_obs_P[i] < 0.05/100){
    out = "both"
  } else {
    if(pdata$Obs_P[i] < 0.05/43 & pdata$wurtz_obs_P[i] > 0.05/100){
      out = "NEO"
    } else {
      if(pdata$Obs_P[i] > 0.05/43 & pdata$wurtz_obs_P[i] < 0.05/100){
        out = "Wurtz"
      } else {
        out = "neither"
      }
    }
  }
  return(out)
})
pdata$sig = factor(pdata$sig, levels = c("NEO","Wurtz","both","neither"))

pcol = RColorBrewer::brewer.pal(4, "Set1")
###
( NEO_Wurtz_scatterplot = pdata %>% ggplot(aes(x = wurtz_obs_beta, y = Obs_beta) ) + 
  geom_abline(slope = 1, intercept = 0, linetype = "dotted", color = "black") +
  geom_vline(xintercept = 0, linetype = "dashed", color = "grey") +
  geom_hline(yintercept = 0, linetype = "dashed", color = "grey") +
  geom_point(aes(color = sig ), size = 3) + 
  # geom_errorbar(aes( xmin  = wurtz_obs_LCI, xmax  = wurtz_obs_UCI, width = 0.15), color = "grey") +
  # geom_errorbar(aes( ymin  = Obs_LCI, ymax  = Obs_UCI, width = 0.15), color = "grey") +
  scale_color_brewer(palette = "Set1") +
  geom_smooth(method = "lm", formula = y~x) +
  labs(x = "Wurtz et al effect estimate", y = "Hughes et al effect estimate", color = "association") +
  annotate("text",x = 0.065, y = -0.09, label = paste0( "Pearson's r = ", round(a$estimate, d = 2)  ) ) +
  geom_text_repel( data = pdata %>% filter(sig == "both" & wurtz_obs_beta > 0 & Obs_beta < 0), 
                  aes(label=outcome), 
                  min.segment.length = 0,
                  box.padding = 0.25, 
                  color = pcol[3], nudge_x = 0.00, nudge_y = -0.02 ) +
  geom_text_repel( data = pdata %>% filter(sig == "both" & wurtz_obs_beta < 0.05 & Obs_beta > 0.05), 
                  aes(label=outcome), 
                  min.segment.length = 0,
                  box.padding = 0.25, 
                  color = pcol[3], nudge_x = -0.001, nudge_y = 0,
                  max.overlaps = 15) +
   geom_text_repel( data = pdata %>% filter(sig == "Wurtz" & wurtz_obs_beta > 0 & Obs_beta < 0), 
                  aes(label=outcome), 
                  min.segment.length = 0,
                  box.padding = 0.25, 
                  color = pcol[2], nudge_x = 0.01, nudge_y = -0.005, 
                  max.overlaps = 15 ) +
   geom_text_repel( data = pdata %>% filter(sig == "Wurtz" & wurtz_obs_beta < 0 ), 
                  aes(label=outcome), 
                  min.segment.length = 0,
                  box.padding = 0.25, 
                  color = pcol[2], nudge_x = -0.01, nudge_y = 0., 
                  max.overlaps = 15) +
  geom_text_repel( data = pdata %>% filter(sig == "NEO"  ), 
                  aes(label=outcome), 
                  min.segment.length = 0,
                  box.padding = 0.25, 
                  color = pcol[1], nudge_x = -0.01, nudge_y = 0, 
                  max.overlaps = 15 ) +
  theme_bw()
)

```


```{r}
f = paste0(resultsdir, "figures/Wurtz_vs_NEO_fasting_cross_sectional.pdf")
pdf(file = f, width = 8, height = 6)
NEO_Wurtz_scatterplot
dev.off()
```


```{r}
table(pdata$sig)
```


## Positive association in Wurtz but no association with negative point estimate here.

```{r}
x = pdata %>% filter(sig == "Wurtz" & wurtz_obs_beta > 0 & Obs_beta < 0)
m = match(x$outcome, study_data$feature_anno$ids)
x = cbind(x, study_data$feature_anno[m,c(8,5,6,7)])
x %>%  kable() %>% kableExtra::kable_classic()
```

```{r}
x = "hdld"
w = grep(x, obs_tsls_est$wNEO$outcome)
obs_tsls_est$wNEO[w,] %>% kable() %>% kableExtra::kable_classic()
```


## POSTPRANDIAL DATA

### correlation between fasting and postprandial

```{r}
fdata = obs_tsls_est$wNEO %>% filter(dietary_state == "fasting") %>% 
  dplyr::select("outcome","Obs_beta","Obs_se","Obs_P", "class", "subclass", "raw.label")
colnames(fdata)[2:4] = paste0("fasting_", colnames(fdata)[2:4])
fdata$metids = metids
##
pdata = obs_tsls_est$wNEO %>% filter(dietary_state == "postprandial") %>% 
  dplyr::select("Obs_beta","Obs_se","Obs_P")
colnames(pdata) = paste0("postprandial_", colnames(pdata))
pdata$metids = metids
##
rdata = obs_tsls_est$wNEO %>% filter(dietary_state == "response") %>% 
  dplyr::select("Obs_beta","Obs_se","Obs_P")
colnames(rdata) = paste0("response_", colnames(rdata))
rdata$metids = metids

df = left_join(fdata, pdata, by = "metids")
df = left_join(df, rdata, by = "metids")
```

## Z-test for difference among estimates

```{r}
z = t( sapply(1:nrow(df), function(i){
  ztest(df$fasting_Obs_beta[i], df$fasting_Obs_se[i], 
        df$postprandial_Obs_beta[i], df$postprandial_Obs_se[i])
}) )

df$ztestP = ifelse( z[,2] < 0.05, "< 0.05", "> 0.05")
df$ztestP = ifelse( z[,2] < 0.05/229, "< 0.05/229", df$ztestP)
df$ztestP = factor(df$ztestP, levels = c("> 0.05","< 0.05","< 0.05/229"))

table(df$ztestP)
```

```{r}
df$association = NULL
for(i in 1:nrow(df)){
  if(df$fasting_Obs_P[i] > 0.05/43 & df$postprandial_Obs_P[i] > 0.05/43){
    out = "neither"
  } else {
    if(df$fasting_Obs_P[i] < 0.05/43 & df$postprandial_Obs_P[i] > 0.05/43){
      out = "fasting"
    } else {
      if(df$fasting_Obs_P[i] > 0.05/43 & df$postprandial_Obs_P[i] < 0.05/43){
        out = "postprandial"
      } else {
        if(df$fasting_Obs_P[i] < 0.05/43 & df$postprandial_Obs_P[i] < 0.05/43){
          out = "both"
        }
      }
    }
  }
  df$association[i] = out
}
df$association = factor(df$association, levels = c("fasting","postprandial","both","neither"))
table(df$association)
```

## observational fasting vs observationl postprandial cor.

```{r}
a = cor.test(df$fasting_Obs_beta, df$postprandial_Obs_beta)
a = paste0("Pearson's r = ", round(a$estimate, d=3) )
a
```


##does response correlate with either fasting or postprandial data?

```{r}
cor.test(fdata$fasting_Obs_beta, rdata$response_Obs_beta)
```


```{r}
cor.test(pdata$postprandial_Obs_beta, rdata$response_Obs_beta)
```



## Fasting association with clearly no assocation in postprandial 

```{r}
df %>% filter(association == "fasting" & postprandial_Obs_P > 0.05) %>% 
  dplyr::select(outcome, fasting_Obs_beta, fasting_Obs_se, fasting_Obs_P, 
                postprandial_Obs_beta, postprandial_Obs_se, postprandial_Obs_P, raw.label) %>%
  arrange( desc(fasting_Obs_beta)) %>%
  kable() %>% kableExtra::kable_classic()
```


## Postprandial association with clearly no assocation in fasting 

```{r}
df %>% filter(association == "postprandial" & fasting_Obs_P > 0.05) %>% 
  dplyr::select(outcome, fasting_Obs_beta, fasting_Obs_P, 
                postprandial_Obs_beta, postprandial_Obs_P, raw.label) %>%
  arrange( desc(postprandial_Obs_beta)) %>%
  kable() %>% kableExtra::kable_classic()
```



```{r}
library(ggrepel)

( fast_post_obs_scatter_plot = df %>% ggplot(aes(x = fasting_Obs_beta, y = postprandial_Obs_beta)) +
  geom_abline(intercept = 0, slope = 1, linetype = "dotted", color = "grey40") + 
  geom_vline(xintercept = 0, linetype = "dashed", color = "grey40") +
  geom_hline(yintercept = 0, linetype = "dashed", color = "grey40") +
  geom_smooth(method = "lm", formula = y~x, color = "black") +
  geom_point(aes(color = association, shape = ztestP), size = 3, alpha = 0.7) +
  scale_color_brewer(palette = "Set1", type = "seq") +
  scale_shape_manual(values = c(19, 15, 17)) +
  theme_bw() +
  labs(x = "fasting cross-sectional estimate", 
       y = "postprandial cross-sectional estimate",
       shape = paste0("Z-test\nbetas differ") ) +
  annotate("text", x = 0.06, y = -0.085, label = a) +
  geom_text_repel( data = df %>% filter(ztestP == "< 0.05/229" & 
                                          postprandial_Obs_beta < fasting_Obs_beta), 
                  aes(label=metids, color = association), 
                  min.segment.length = 0,
                  box.padding = 0.25, 
                  nudge_x = 0.005, nudge_y = -0.001 ) +
  geom_text_repel( data = df %>% filter(ztestP == "< 0.05/229" & 
                                          postprandial_Obs_beta > fasting_Obs_beta), 
                  aes(label=metids, color = association), 
                  min.segment.length = 0,
                  box.padding = 0.25, 
                  nudge_x = -0.008, nudge_y = 0.01 )
  )
```

```{r}
f = paste0(resultsdir, "figures/wNEO_fasting_postprandial_cross_sectional.pdf")
pdf(file = f, width = 8, height = 6)
fast_post_obs_scatter_plot
dev.off()
```


```{r}
table(df$association)
```

```{r}
table(df$ztestP)
```

```{r}
table(df$ztestP, df$association)
```

```{r}
table(unlist( df %>% filter(ztestP  == "< 0.05/229") %>% dplyr::select(class) ) )
```

```{r}
df %>% filter(ztestP  == "< 0.05/229") %>% arrange(desc(fasting_Obs_beta)) %>%
  kable() %>% kableExtra::kable_classic()
```



```{r, fig.width = 7.5, fig.height = 10}
library(patchwork)
male_female_scatter_plot / NEO_Wurtz_scatterplot / fast_post_obs_scatter_plot +
   plot_annotation(tag_levels = 'A')
```


```{r}
f = paste0(resultsdir, "figures/wNEO_cross_sectional_multiplot.pdf")
pdf(file = f, width = 8, height = 12)
male_female_scatter_plot / NEO_Wurtz_scatterplot / fast_post_obs_scatter_plot +
   plot_annotation(tag_levels = 'A')
dev.off()
```



## POSTPANDRIAL LIPIOPROTEINS ASSOCIATION PATTERN

```{r, fig.width = 10, fig.height = 8, error=FALSE, warning=FALSE}

## Function to extract needed data
wdata = lipoprotein_wdata(wd = obs_tsls_est$wNEO, define_Dstate = "postprandial")

## plot order of subclasses
wdata$trait = factor(wdata$trait , levels = c("concentration", "total_lipids", 
                                              "phospholipids", "free_cholesterol", 
                                              "total_cholesterol", "cholesterol_esters", 
                                              "triglycerides",
                                              "phospholipids/total_lipids", "free_cholesterol/total_lipids", 
                                              "total_cholesterol/total_lipids" , 
                                              "cholesterol_esters/total_lipids", 
                                              "triglycerides/total_lipids")[12:1] )

###
x = RColorBrewer::brewer.pal(8,"Set1")
pcol = c( rep(x[3],6) , x[4] , rep(x[1], 3), rep(x[2],4) ) 
###
plotA = wdata %>% ggplot(aes(x = subclass, y = trait, fill = Obs_beta)) +
  geom_tile() + 
  geom_text(data = subset(wdata, Obs_P <= 0.05/43), aes(label =  round(Obs_beta, d = 4) ), size = 2.4 ) +
  # geom_text(data = subset(wdata, MR_P <= 0.05/43), aes(label =  paste0(round(MR_beta, d = 4), "*" ) ), size = 2.4 ) +
  scale_fill_gradient2(low = "red", mid = "white", high = "blue") +
  theme(axis.text.x = element_text(angle = 35, vjust = 1, hjust=1, colour = pcol)) +
  labs(x = "lipoprotein subclass", y = "trait", fill = "beta") +
  facet_grid(value ~ ., scales = "free_y", space = "free_y")

plotB = wdata  %>% 
  filter(!grepl("/",trait)) %>% 
  ggplot(aes(x = as.numeric(subclass), y = Obs_beta, color = trait)) +
  geom_hline(yintercept = 0, size = 1, linetype = "dashed", color = "grey30") +
  geom_point(size = 4, alpha = 0.6) +
  geom_line( aes(color = trait) ) +
  labs(x = "lipoprotein subclass", y = "beta", color = "trait") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 40, vjust = 1, hjust=1, color = pcol)) + 
  scale_x_continuous(name = "lipoprotein subclass", breaks = 1:14, labels = levels(wdata$subclass) )

ggpubr::ggarrange(plotA, plotB, ncol = 1)  

```


## save postprandial lipoproteins beta profile

```{r}
f = paste0(resultsdir, "figures/wNEO_postprandial_lipo_Obs_profile.pdf")
pdf(file = f, width = 10, height = 8)
ggpubr::ggarrange(plotA, plotB, ncol = 1)  
dev.off()
```

## Are the lipoprotein lipid effect estiamtes larger in the postprandial state?

```{r}
temp = df %>% filter(class == "Lipoprotein subclasses") %>% 
  mutate( post_beta_larger_than_fast = ifelse( abs(postprandial_Obs_beta) - abs(fasting_Obs_beta) > 0, "yes","no" ) )
                  
table(temp$post_beta_larger_than_fast)
```

```{r}
temp %>% filter(post_beta_larger_than_fast == "no") %>% dplyr::select(metids, raw.label, fasting_Obs_beta, postprandial_Obs_beta ) %>% kable() %>% kableExtra::kable_classic()
```

```{r}
delta = unlist( df %>% filter(class == "Lipoprotein subclasses") %>% 
  summarize( delta = abs(postprandial_Obs_beta) - abs(fasting_Obs_beta)  ) )
names(delta) = temp$metids   
summary(delta)

t.test(temp$fasting_Obs_beta, temp$postprandial_Obs_beta)
```


```{r}
df %>% filter(metids %in% c("apoa1", "apob", "apobapoa1")) %>% 
  kable() %>% kableExtra::kable_classic()
```



## RESPONSE

## Assocations

```{r}
rsig = obs_tsls_est$wNEO %>% 
  filter(dietary_state == "response", Obs_P < 0.05/43) %>% 
  dplyr::select(outcome, Obs_devexp_by_exposure,
                Obs_beta,Obs_se,Obs_P, class, subclass, raw.label) %>% arrange(Obs_P) 

rsig %>% kable() %>%kableExtra::kable_classic()
```

## count positive and inverse associations

```{r}
rsig %>% summarise(pos = sum(Obs_beta > 0), neg = sum(Obs_beta<0))
```

## Amino Acids

```{r}
rsig %>% filter(class == "Amino acids") %>% arrange(Obs_P) %>% 
  kable() %>%kableExtra::kable_classic()
```

```{r}
rsig %>% filter(outcome %in% c("tyr_r", "ala_r") )
```


## effect estiamtes without a RNT

```{r}
tr = c("tyr_r", "ala_r")
##
x = t(sapply( tr, function(x){
  obsmr( wdata = study_data$working_data,
            outcome = x,
            exposure = "bmi",
            instrument = "Yengo_Weighted_GRS",
            covariates = c("visit_date","subpop","age","sex"),
            weights = "pweight",
            rnt_outcome = FALSE,
            outlier_method = "iqr",
            outlier_cutoff = 10,
            messages = TRUE)
}))
x = as.data.frame(x)
x %>% kable() %>% kableExtra::kable_classic()
```


## A response example : ALAININE

## Estimate observational difference in ALAININE abundance in fasted and postprandial states

```{r}
cat(paste0("postprandial mean\n"))
mean(mydata$ala_p, na.rm = TRUE)
cat(paste0("fasting mean\n"))
mean(mydata$ala_f, na.rm = TRUE)

cat(paste0("Paired t.test\n"))
t.test(mydata$ala_p, mydata$ala_f, paired = TRUE, alternative = "two.sided")$p.value

cat(paste0("linear model\n"))
fit = lm(ala_p ~ ala_f, data = mydata)
summary(fit)
```


```{r, fig.width = 10, fig.height = 4}
plot_data = na.omit( mydata[, c("sex", "age", "ala_f","ala_p","ala_r","bmi", "bmi_cat")] )
plot_data = as.data.frame(plot_data)

fit = lm(ala_r ~ sex + age + bmi  , data = plot_data)

library(mgcv)
fit2 = gam(ala_r ~ sex + age + s(bmi), data = plot_data)

pcol = RColorBrewer::brewer.pal(8, "Set1")

#################
p1 = plot_data %>% ggplot(aes(x = ala_f, y = ala_p)) +
  geom_point( color = "grey80" , alpha = .50, size = 3) +
  geom_abline(intercept = 0, slope = 1, linetype = "dotted", size = 1.25) +
  #geom_vline(xintercept = c(0.3, 0.4, 0.5), linetype = "dotted") +
  #geom_hline(yintercept = c(0.3, 0.4, 0.5, 0.6), linetype = "dotted") +
  geom_smooth( method = lm, formula = y~x-1, aes(color = bmi_cat)  ) +
  scale_color_brewer(palette="Set1") +
  theme_bw() +
  labs(color = "BMI category", x = "fasting alanine", y = "postprandial alanine")
  
###############
p2 = plot_data %>% ggplot(aes(x = bmi, y = ala_r)) +
  #geom_point( color = "grey80" , alpha = .50, size = 3) +
  geom_smooth( method = lm, formula = y~x, color = "grey40", fill = "grey", linetype = "dashed") +
  geom_smooth( method = gam, formula = y~s(x), 
               color = "black", fill = "black",
               size = 3, se = FALSE ) +
  geom_smooth( method = lm, formula = y~x, aes(color = bmi_cat)  ) +
  scale_color_brewer(palette="Set1") +
  labs(color = "BMI category", x = "BMI", y = "alanine response") +
  theme_bw()

# (p = ggarrange(p1, p2, nrow = 1, common.legend = TRUE) )
p1 + p2 + plot_annotation( tag_levels = 'A' ) + plot_layout(guides = "collect") & theme(legend.position = "bottom") 

```

## Save alanine example to file

```{r}
f = paste0(resultsdir, "figures/obs_alanine_response_v2.pdf")
pdf(file = f, width = 10, height = 4)
p1 + p2 + plot_annotation( tag_levels = 'A' ) + plot_layout(guides = "collect") & theme(legend.position = "bottom") 
dev.off()
```


## Top Response postive associations

```{r}
rsig = obs_tsls_est$wNEO %>% 
  filter(dietary_state == "response", Obs_P < 0.05/43) %>% 
  dplyr::select(outcome, Obs_devexp_by_exposure,
                Obs_beta,Obs_se,Obs_P, class, subclass, raw.label) %>% arrange(Obs_P) 

rsig %>% filter(Obs_beta > 0) %>% kable() %>%kableExtra::kable_classic()
```



## A response example : xlhdltg_r

## Estimate observational difference in xlhdltg_r abundance in fasted and postprandial states

```{r}
cat(paste0("postprandial mean\n"))
mean(mydata$xlhdltg_p, na.rm = TRUE)
cat(paste0("fasting mean\n"))
mean(mydata$xlhdltg_f, na.rm = TRUE)

cat(paste0("Paired t.test\n"))
t.test(mydata$xlhdltg_p, mydata$xlhdltg_f, paired = TRUE, alternative = "two.sided")$p.value

cat(paste0("linear model\n"))
fit = lm(xlhdltg_p ~ xlhdltg_f, data = mydata)
summary(fit)
```


```{r, fig.width = 10, fig.height = 4}
plot_data = na.omit( mydata[, c("sex", "age", "xlhdltg_f","xlhdltg_p","xlhdltg_r","bmi", "bmi_cat")] )
plot_data = as.data.frame(plot_data)

fit = lm(xlhdltg_r ~ sex + age + bmi  , data = plot_data)

library(mgcv)
fit2 = gam(xlhdltg_r ~ sex + age + s(bmi), data = plot_data)

pcol = RColorBrewer::brewer.pal(8, "Set1")

#################
p1 = plot_data %>% ggplot(aes(x = xlhdltg_f, y = xlhdltg_p)) +
  #geom_point( color = "grey80" , alpha = .50, size = 3) +
  geom_abline(intercept = 0, slope = 1, linetype = "dotted", size = 1.25) +
  #geom_vline(xintercept = c(0.3, 0.4, 0.5), linetype = "dotted") +
  #geom_hline(yintercept = c(0.3, 0.4, 0.5, 0.6), linetype = "dotted") +
  geom_smooth( method = lm, formula = y~x-1, aes(color = bmi_cat)  ) +
  scale_color_brewer(palette="Set1") +
  theme_bw() +
  labs(color = "BMI category", x = "fasting xlhdltg", y = "postprandial xlhdltg")
  
###############
p2 = plot_data %>% ggplot(aes(x = bmi, y = xlhdltg_r)) +
  #geom_point( color = "grey80" , alpha = .50, size = 3) +
  geom_smooth( method = lm, formula = y~x, color = "grey40", fill = "grey", linetype = "dashed") +
  geom_smooth( method = gam, formula = y~s(x), 
               color = "black", fill = "black",
               size = 3, se = FALSE ) +
  geom_smooth( method = lm, formula = y~x, aes(color = bmi_cat)  ) +
  scale_color_brewer(palette="Set1") +
  labs(color = "BMI category", x = "BMI", y = "xlhdltg response") +
  theme_bw()

## Plot 
p1 + p2 + plot_annotation( tag_levels = 'A' ) + plot_layout(guides = "collect") & theme(legend.position = "bottom") 


```


## RESPONSE LIPIOPROTEINS ASSOCIATION PATTERN

```{r, fig.width = 10, fig.height = 8, error=FALSE, warning=FALSE}

## Function to extract needed data
wdata = lipoprotein_wdata(wd = obs_tsls_est$wNEO, define_Dstate = "response")

## plot order of subclasses
wdata$trait = factor(wdata$trait , levels = c("concentration", "total_lipids", 
                                              "phospholipids", "free_cholesterol", 
                                              "total_cholesterol", "cholesterol_esters", 
                                              "triglycerides",
                                              "phospholipids/total_lipids", "free_cholesterol/total_lipids", 
                                              "total_cholesterol/total_lipids" , 
                                              "cholesterol_esters/total_lipids", 
                                              "triglycerides/total_lipids")[12:1] )

###
x = RColorBrewer::brewer.pal(8,"Set1")
pcol = c( rep(x[3],6) , x[4] , rep(x[1], 3), rep(x[2],4) ) 
###
plotA = wdata %>% ggplot(aes(x = subclass, y = trait, fill = Obs_beta)) +
  geom_tile() + 
  geom_text(data = subset(wdata, Obs_P <= 0.05/43), aes(label =  round(Obs_beta, d = 4) ), size = 2.4 ) +
  # geom_text(data = subset(wdata, MR_P <= 0.05/43), aes(label =  paste0(round(MR_beta, d = 4), "*" ) ), size = 2.4 ) +
  scale_fill_gradient2(low = "red", mid = "white", high = "blue") +
  theme(axis.text.x = element_text(angle = 35, vjust = 1, hjust=1, colour = pcol)) +
  labs(x = "lipoprotein subclass", y = "trait", fill = "beta") +
  facet_grid(value ~ ., scales = "free_y", space = "free_y")

plotB = wdata  %>% 
  filter(!grepl("/",trait)) %>% 
  ggplot(aes(x = as.numeric(subclass), y = Obs_beta, color = trait)) +
  geom_hline(yintercept = 0, size = 1, linetype = "dashed", color = "grey30") +
  geom_point(size = 4, alpha = 0.6) +
  geom_line( aes(color = trait) ) +
  labs(x = "lipoprotein subclass", y = "beta", color = "trait") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 40, vjust = 1, hjust=1, color = pcol)) + 
  scale_x_continuous(name = "lipoprotein subclass", breaks = 1:14, labels = levels(wdata$subclass) )

ggpubr::ggarrange(plotA, plotB, ncol = 1)  

```

## save response lipoproteins beta profile

```{r}
f = paste0(resultsdir, "figures/wNEO_response_lipo_Obs_profile.pdf")
pdf(file = f, width = 10, height = 8)
ggpubr::ggarrange(plotA, plotB, ncol = 1)  
dev.off()
```


```{r}
rsig %>% filter(!class %in% c("Amino acids", "Lipoprotein subclasses", "Lipoprotein subclasses ratios")) %>% dplyr::select(Obs_beta, Obs_se, Obs_P, class, raw.label) %>%
  kable() %>% kableExtra::kable_classic()
```



















## Observational effect estimates

```{r , fig.width = 9, fig.height = 11}
k = c( "Lipoprotein particle size",
       "Glycerides and phospholipids",
       "Cholesterol" , "Apolipoproteins",
       "Fatty acids",
       "Fatty acids ratios",
       "Amino acids", 
      "Glycolysis related metabolites", "Ketone bodies", "Inflammation")

wdata = obs_tsls_est$wNEO %>% filter( class %in% k  ) %>%
  dplyr::select( outcome, dietary_state, class, raw.label, Obs_beta, Obs_P)
## edit outcome names
wdata$outcome = gsub("_f","",wdata$outcome);wdata$outcome = gsub("_p","",wdata$outcome);wdata$outcome = gsub("_r","",wdata$outcome)
## edit plot order with factor level
o = order( wdata %>% filter(dietary_state == "fasting") %>% dplyr::select(Obs_beta) )
wdata$outcome = factor(wdata$outcome, levels = unique(wdata$outcome)[o] )
## Dietary State plot order
wdata$dietary_state = factor(wdata$dietary_state, levels = unique(wdata$dietary_state) )
## Class plot order
wdata$class = factor(wdata$class, levels = unique(wdata$class) )


#########3
plotA = wdata %>% ggplot(aes(y = outcome, x = dietary_state, fill = Obs_beta)) +
  geom_tile() + 
  geom_text(data = subset(wdata, Obs_P < 0.05/43), aes(label =  round(Obs_beta,d = 4)  ), size = 2.75 ) +
  scale_fill_gradient2(low = "red", mid = "white", high = "blue") +
  theme(axis.text.x = element_text(angle = 35, vjust = 1, hjust=1)) +
  labs(x = "dietary state", y = "outcome") +
  facet_grid(class ~ ., scales = "free_y", space = "free_y") +
  theme(strip.text.y = element_text(angle = 0))

plotA

```


## save beta profiles

```{r}
f = paste0(resultsdir, "figures/Obs_Beta_AllDietaryState_profiles_v2.pdf")
pdf(file = f, width = 9, height = 11)
plotA
dev.off()
```


## Plot lipoprotein subclasses observational betas


```{r , fig.width = 9, fig.height = 16}
k = c( "Lipoprotein subclasses")

wdata = obs_tsls_est$wNEO %>% filter( class %in% k ) %>%
  dplyr::select( outcome, dietary_state, class, subclass, raw.label, Obs_beta, Obs_P)
## edit outcome names
wdata$outcome = gsub("_f","",wdata$outcome);wdata$outcome = gsub("_p","",wdata$outcome);wdata$outcome = gsub("_r","",wdata$outcome)
## edit plot order with factor level
o = order( wdata %>% filter(dietary_state == "fasting") %>% dplyr::select(Obs_beta) )
wdata$outcome = factor(wdata$outcome, levels = unique(wdata$outcome)[o] )
## Dietary State plot order
wdata$dietary_state = factor(wdata$dietary_state, levels = unique(wdata$dietary_state) )
## Class plot order
wdata$subclass = factor(wdata$subclass, levels = unique(wdata$subclass) )


#########3
plotA = wdata %>% ggplot(aes(y = outcome, x = dietary_state, fill = Obs_beta)) +
  geom_tile() + 
  geom_text(data = subset(wdata, Obs_P < 0.05/43), aes(label =  round(Obs_beta,d = 4)  ), size = 2.75 ) +
  scale_fill_gradient2(low = "red", mid = "white", high = "blue") +
  theme(axis.text.x = element_text(angle = 35, vjust = 1, hjust=1)) +
  labs(x = "dietary state", y = "outcome") +
  facet_grid(subclass ~ ., scales = "free_y", space = "free_y") +
  theme(strip.text.y = element_text(angle = 0))

plotA

```



## save beta profiles

```{r}
f = paste0(resultsdir, "figures/Obs_Beta_AllDietaryState_Lipoprotein_profiles_v2.pdf")
pdf(file = f, width = 9, height = 11)
plotA
dev.off()
```



## Plot lipoprotein subclasses observational betas


```{r , fig.width = 9, fig.height = 16}
k = c( "Lipoprotein subclasses ratios")

wdata = obs_tsls_est$wNEO %>% filter( class %in% k ) %>%
  dplyr::select( outcome, dietary_state, class, subclass, raw.label, Obs_beta, Obs_P)
## edit outcome names
wdata$outcome = gsub("_f","",wdata$outcome);wdata$outcome = gsub("_p","",wdata$outcome);wdata$outcome = gsub("_r","",wdata$outcome)
## edit plot order with factor level
o = order( wdata %>% filter(dietary_state == "fasting") %>% dplyr::select(Obs_beta) )
wdata$outcome = factor(wdata$outcome, levels = unique(wdata$outcome)[o] )
## Dietary State plot order
wdata$dietary_state = factor(wdata$dietary_state, levels = unique(wdata$dietary_state) )
## Class plot order
wdata$subclass = factor(wdata$subclass, levels = unique(wdata$subclass) )


#########3
plotA = wdata %>% ggplot(aes(y = outcome, x = dietary_state, fill = Obs_beta)) +
  geom_tile() + 
  geom_text(data = subset(wdata, Obs_P < 0.05/43), aes(label =  round(Obs_beta,d = 4)  ), size = 2.75 ) +
  scale_fill_gradient2(low = "red", mid = "white", high = "blue") +
  theme(axis.text.x = element_text(angle = 35, vjust = 1, hjust=1)) +
  labs(x = "dietary state", y = "outcome") +
  facet_grid(subclass ~ ., scales = "free_y", space = "free_y") +
  theme(strip.text.y = element_text(angle = 0))

plotA

```



## save beta profiles

```{r}
f = paste0(resultsdir, "figures/Obs_Beta_AllDietaryState_Lipoprotein_Ratio_profiles_v2.pdf")
pdf(file = f, width = 9, height = 11)
plotA
dev.off()
```



## Positive and Negative Subclass counts

```{r}
fast_neg  = obs_tsls_est$wNEO %>% 
  filter( Obs_beta < 0 & dietary_state == "fasting" & !grepl("pct", outcome) & Obs_P < 0.05/43 ) %>%
  dplyr::select(outcome, Obs_beta, Obs_se, Obs_P, raw.label, class, subclass)
###
table(fast_neg$subclass)


fast_pos  = obs_tsls_est$wNEO %>% 
  filter( Obs_beta > 0 & dietary_state == "fasting" & !grepl("pct", outcome) & Obs_P < 0.05/43 ) %>%
  dplyr::select(outcome, Obs_beta, Obs_se, Obs_P, raw.label, class, subclass)
###
table(fast_pos$subclass)


```

```{r}
fast_neg %>% arrange(Obs_P) %>% kable() %>% kable_classic()
```



## Response observational effect estimates 

## POSITIVE Response observational associations

```{r}
pos_mytable = obs_tsls_est$wNEO %>% filter(dietary_state == "response" & Obs_beta > 0 & Obs_P < 0.05/43) %>% 
  arrange( desc( abs( Obs_beta ) ) ) %>%
  dplyr::select(outcome, raw.label, class, Obs_beta, Obs_se, Obs_P) 

pos_mytable %>% kable %>% kable_classic()
```

## NEGATIVE Response observational associations

```{r}
neg_mytable = obs_tsls_est$wNEO %>% filter(dietary_state == "response" & Obs_beta < 0 & Obs_P < 0.05/43) %>% 
  arrange( desc( abs( Obs_beta ) ) ) %>%
  dplyr::select(outcome, raw.label, class, Obs_beta, Obs_se, Obs_P) 

neg_mytable %>% kable %>% kable_classic()
```

## Save table to file

```{r}
mytable = rbind(neg_mytable, pos_mytable)
f = paste0(resultsdir, "tables/wNEO_Obs_RESPONSE_associations_v2.txt")
write.table(mytable, file = f, row.names = FALSE, col.names = TRUE, sep = "\t", quote = TRUE)

```


## how many metabolite response traits associate with BMI positively and inversely? 

```{r}
df %>% filter(dietary_state == "response" & Obs_P < 0.05/43) %>% 
  summarise(neg = sum(Obs_beta < 0), pos = sum(Obs_beta > 0))
```

## A response example : ALAININE

```{r, fig.width = 10, fig.height = 4}
plot_data = na.omit( mydata[, c("sex", "age", "ala_f","ala_p","ala_r","bmi")] )
w = id_outliers(plot_data$ala_r)
if(length(w)>0){ plot_data = plot_data[-w,] }

##
plot_data = as.data.frame(plot_data)
plot_data$bmi_class = "healthy weight"
w = which( plot_data$bmi > 25  )
plot_data$bmi_class[w] = "overweight"
w = which( plot_data$bmi > 30 )
plot_data$bmi_class[w] = "obese"
w = which( plot_data$bmi > 40 )
plot_data$bmi_class[w] = "severely obese"

plot_data$bmi_class = factor(plot_data$bmi_class, 
                             levels = c("healthy weight","overweight","obese", "severely obese") )

fit = lm(ala_r ~ sex + age + bmi  , data = plot_data)
library(mgcv)
fit2 = gam(ala_r ~ sex + age + s(bmi), data = plot_data)

pcol = RColorBrewer::brewer.pal(8, "Set1")

#################
p1 = plot_data %>% ggplot(aes(x = ala_f, y = ala_p)) +
  geom_point( color = "grey80" , alpha = .50, size = 3) +
  geom_abline(intercept = 0, slope = 1, linetype = "dotted", size = 1.25) +
  #geom_vline(xintercept = c(0.3, 0.4, 0.5), linetype = "dotted") +
  #geom_hline(yintercept = c(0.3, 0.4, 0.5, 0.6), linetype = "dotted") +
  geom_smooth( method = lm, formula = y~x-1, aes(color = bmi_class)  ) +
  scale_color_brewer(palette="Set1") +
  theme_bw() +
  labs(color = "BMI category", x = "fasting alanine", y = "postprandial alanine")
  
###############
p2 = plot_data %>% ggplot(aes(x = bmi, y = ala_r)) +
  geom_point( color = "grey80" , alpha = .50, size = 3) +
  geom_smooth( method = lm, formula = y~x, color = "grey40", fill = "grey", linetype = "dashed") +
  geom_smooth( method = gam, formula = y~s(x), 
               color = "black", fill = "black",
               size = 3, se = FALSE ) +
  geom_smooth( method = lm, formula = y~x, aes(color = bmi_class)  ) +
  scale_color_brewer(palette="Set1") +
  labs(color = "BMI category", x = "BMI", y = "alanine response") +
  theme_bw()

p = ggarrange(p1, p2, nrow = 1, common.legend = TRUE)
p


```

## Save alanine example to file

```{r}
f = paste0(resultsdir, "figures/obs_alanine_response.pdf")
pdf(file = f, width = 10, height = 4)
p
dev.off()
```


## A response example : LEUCINE

```{r, fig.width = 12, fig.height = 9}
p = bmi_metabolite_plot(data = mydata, metabolite = "leu", metabolite_name = "leucine")
p
```

## mean change in amino acids between fasted and postprandial states

```{r}
all_mean = mydata  %>% summarize(mean_bmi = mean(bmi, na.rm = TRUE),
                                             mean_ala_delta = mean(ala_p - ala_f, na.rm = TRUE),
                                             mean_gln_delta = mean(gln_p - gln_f, na.rm = TRUE),
                                             mean_his_delta = mean(his_p - his_f, na.rm = TRUE),
                                             mean_phe_delta = mean(phe_p - phe_f, na.rm = TRUE),
                                             mean_tyr_delta = mean(tyr_p - tyr_f, na.rm = TRUE),
                                             mean_ile_delta = mean(ile_p - ile_f, na.rm = TRUE),
                                             mean_leu_delta = mean(leu_p - leu_f, na.rm = TRUE),
                                             mean_val_delta = mean(val_p - val_f, na.rm = TRUE)
                                             )
all_mean = cbind( bmi_class = "all", all_mean)
all_mean
```


```{r}
mydata$bmi_class = "healthy weight"
w = which( mydata$bmi > 25  )
mydata$bmi_class[w] = "overweight"
w = which( mydata$bmi > 30 )
mydata$bmi_class[w] = "obese"
w = which( mydata$bmi > 40 )
mydata$bmi_class[w] = "severely obese"
mydata$bmi_class = factor(mydata$bmi_class, levels = c("healthy weight","overweight","obese","severely obese"))

########
aas = c("ala","gln","his","phe","tyr", "ile", "leu", "val")
####
mytable = mydata %>% group_by(bmi_class) %>% summarize(mean_bmi = mean(bmi, na.rm = TRUE),
                                             mean_ala_delta = mean(ala_p - ala_f, na.rm = TRUE),
                                             mean_gln_delta = mean(gln_p - gln_f, na.rm = TRUE),
                                             mean_his_delta = mean(his_p - his_f, na.rm = TRUE),
                                             mean_phe_delta = mean(phe_p - phe_f, na.rm = TRUE),
                                             mean_tyr_delta = mean(tyr_p - tyr_f, na.rm = TRUE),
                                             mean_ile_delta = mean(ile_p - ile_f, na.rm = TRUE),
                                             mean_leu_delta = mean(leu_p - leu_f, na.rm = TRUE),
                                             mean_val_delta = mean(val_p - val_f, na.rm = TRUE)
                                             )
mytable = rbind(all_mean, mytable)
mytable %>% kable() %>% kable_classic()
```


```{r}
par(mfrow = c(2, 4))
for(i in 3:10){
  par(new = FALSE)
  barplot(mytable[,i], 
          col = c("black", RColorBrewer::brewer.pal(4, "Set1") ),
          main = colnames(mytable)[i] )  
}

```


```{r}
df %>% filter(outcome %in% c( "tyr_r", "ala_r") ) %>% kable() %>% kableExtra::kable_classic()
```





## Estimate observational difference in ALANINE abundance in fasted and postprandial states

```{r}
cat(paste0("postprandial mean\n"))
mean(mydata$ala_p, na.rm = TRUE)
cat(paste0("fasting mean\n"))
mean(mydata$ala_f, na.rm = TRUE)

cat(paste0("Paired t.test\n"))
t.test(mydata$ala_p, mydata$tyr_f, paired = TRUE, alternative = "two.sided")$p.value

cat(paste0("linear model\n"))
fit = lm(ala_p ~ ala_f, data = mydata)
summary(fit)
```



## A response example : TYROSINE

```{r, fig.width = 10, fig.height = 4}
plot_data = na.omit( mydata[, c("sex", "age", "tyr_f","tyr_p","tyr_r","bmi")] )
plot_data = as.data.frame(plot_data)
plot_data$bmi_class = "healthy weight"
w = which( plot_data$bmi > 25  )
plot_data$bmi_class[w] = "overweight"
w = which( plot_data$bmi > 30 )
plot_data$bmi_class[w] = "obese"
w = which( plot_data$bmi > 40 )
plot_data$bmi_class[w] = "severely obese"

plot_data$bmi_class = factor(plot_data$bmi_class, 
                             levels = c("healthy weight","overweight","obese", "severely obese") )

fit = lm(tyr_r ~ sex + age + bmi  , data = plot_data)
library(mgcv)
fit2 = gam(tyr_r ~ sex + age + s(bmi), data = plot_data)

pcol = RColorBrewer::brewer.pal(8, "Set1")

#################
p1 = plot_data %>% ggplot(aes(x = tyr_f, y = tyr_p)) +
  geom_point( color = "grey80" , alpha = .50, size = 3) +
  geom_abline(intercept = 0, slope = 1, linetype = "dotted", size = 1.25) +
  geom_smooth( method = lm, formula = y~x-1, aes(color = bmi_class)  ) +
  scale_color_brewer(palette="Set1") +
  theme_bw() +
  labs(color = "BMI category", x = "fasting tyrosine", y = "postprandial tyrosine")
  
###############
p2 = plot_data %>% ggplot(aes(x = bmi, y = tyr_r)) +
  geom_point( color = "grey80" , alpha = .50, size = 3) +
  geom_smooth( method = lm, formula = y~x, color = "grey40", fill = "grey", linetype = "dashed") +
  geom_smooth( method = gam, formula = y~s(x), 
               color = "black", fill = "black",
               size = 3, se = FALSE ) +
  geom_smooth( method = lm, formula = y~x, aes(color = bmi_class)  ) +
  scale_color_brewer(palette="Set1") +
  labs(color = "BMI category", x = "BMI", y = "tyrosine response") +
  theme_bw()

p = ggarrange(p1, p2, nrow = 1, common.legend = TRUE)
p


```

## Save tyorsine example to file

```{r}
f = paste0(resultsdir, "figures/obs_tyrpsine_response.pdf")
pdf(file = f, width = 10, height = 4)
p
dev.off()
```

## Estimate observational difference in TYROSINE abundance in fasted and postprandial states

```{r}
cat(paste0("postprandial mean\n"))
mean(mydata$tyr_p, na.rm = TRUE)
cat(paste0("fasting mean\n"))
mean(mydata$tyr_f, na.rm = TRUE)

cat(paste0("Paired t.test\n"))
t.test(mydata$tyr_p, mydata$tyr_f, paired = TRUE, alternative = "two.sided")$p.value

cat(paste0("linear model\n"))
fit = lm(tyr_p ~ tyr_f, data = mydata)
summary(fit)
```






## Correlation between fasting and postprandial observational estimates

```{r}
library(ggrepel)
###
f = grep( "_f", obs_tsls_est$wNEO$outcome )
p = grep( "_p", obs_tsls_est$wNEO$outcome )
####
temp = data.frame( outcome = obs_tsls_est$wNEO$outcome[f],
                   fasting = obs_tsls_est$wNEO$Obs_beta[f],
                   fasting_se = obs_tsls_est$wNEO$Obs_se[f],
                   fasting_P = obs_tsls_est$wNEO$Obs_P[f],
                   postprandial = obs_tsls_est$wNEO$Obs_beta[p],
                   postprandial_se = obs_tsls_est$wNEO$Obs_se[p],
                   postprandial_P = obs_tsls_est$wNEO$Obs_P[p] )
temp$outcome = gsub("_f","",temp$outcome )

PearsonR = cor.test(temp$fasting, temp$postprandial)$estimate
PearsonR = formatC(PearsonR, digits = 3)


#### difference between point estimates
z = t( sapply(1:nrow(temp), function(i){
  ztest(beta1 = temp[i,2], se1 = temp[i,3] , beta2 = temp[i,5] , se2 = temp[i,6] )  
}) )
temp$ztest_P = z[,2]
w = which( z[,2]< 0.05 )
q = which( z[,2]< 0.05/nrow(z) )
temp$sig_dif = "P>0.05"
temp$sig_dif[w] = "P<0.05"
temp$sig_dif[q] = "P<0.05/229"
temp$sig_dif = factor(temp$sig_dif, levels = c("P<0.05/229", "P<0.05" ,"P>0.05" ) )
###
pcol = RColorBrewer::brewer.pal(8, "Set1")
###
p = temp %>% ggplot(aes(x = fasting, y = postprandial)) +
  geom_smooth( method = "lm", formula = y ~ x , color = "black", size = 3) +
  geom_abline(intercept = 0, slope = 1, color = "red") +
  geom_vline(xintercept = 0, linetype = "dashed", color = "grey50") +
  geom_hline(yintercept = 0, linetype = "dashed", color = "grey50") +
  geom_point( aes(color = as.factor(sig_dif) ), size = 4, alpha = 0.5) +
  geom_text_repel( data = temp %>% filter(sig_dif == "P<0.05/229" & fasting > postprandial), 
                  aes(label=outcome), 
                  min.segment.length = 0,
                  box.padding = 0.25, 
                  color = pcol[1], nudge_x = 0.015 ) +
  geom_text_repel( data = temp %>% filter(sig_dif == "P<0.05/229" & fasting < postprandial), 
                  aes(label=outcome), 
                  min.segment.length = 0,
                  box.padding = 0.25, 
                  color = pcol[1], nudge_x = -0.02, nudge_y = 0.01 ) +
  scale_color_brewer(palette="Set1") +
  labs(color = paste0("estimates diff."), 
       subtitle = paste0( "Pearson's r = ", PearsonR ) ) +
  theme_bw()


p 
```

## save fasting postprandial scatter plot to file

```{r}
f = paste0(resultsdir, "figures/obs_fasting_postprandial_cor.pdf")
pdf(file = f, width = 6, height = 4)
p
dev.off()
```

## count how many point estimates differ between dietary states

```{r}
table(temp$sig_dif)
```

## which metabolites are different? and order them

```{r}
q = temp %>% filter(sig_dif == "P<0.05/229") %>% arrange(ztest_P)
n = q$outcome
###
m = match(n, study_data$feature_anno$metabolite)

study_data$feature_anno[m, c("metabolite", "label","class","subclass")] %>% kbl() %>%
  kable_classic(full_width = F, html_font = "Cambria")
```

## Which metabolites association calls are different between dietary states

```{r}
################################
## identify those associated at fasting but not postprandial
################################
fnotp = temp %>% 
  filter(fasting_P < 0.05/43 & postprandial_P > 0.05/43) %>% 
  arrange( desc(fasting) ) %>%
  dplyr::select(outcome)
fnotp = fnotp$outcome
##
length(fnotp) ## 11

################################
## identify those associated at postpranial but not fasting
################################
pnotf = temp %>% 
  filter(fasting_P > 0.05/43 & postprandial_P < 0.05/43) %>% 
  arrange( desc(postprandial) ) %>%
  dplyr::select(outcome)
pnotf = pnotf$outcome
##
length(pnotf) ## 21

################################
## identify those associated with both postpranial and fasting
## but have a different estimate
################################
pf_dif = temp %>% 
  filter(fasting_P < 0.05/43 & postprandial_P < 0.05/43 & ztest_P < 0.05/229 ) %>% 
  arrange( desc(fasting) ) %>%
  dplyr::select(outcome)
pf_dif = pf_dif$outcome
##
length(pf_dif)  # 8
#######
metnames = c(fnotp, pnotf, pf_dif )

##
w = match(metnames, temp$outcome)
assocition_dif = temp[ w, ]
assocition_dif %>% kbl() %>%
  kable_classic(full_width = F, html_font = "Cambria")

```



```{r}
long_data = c()
for(i in 1:length(obs_tsls_est_long)){
  long_data = rbind(long_data, obs_tsls_est_long[[i]] )
}
###
# table(long_data$analysis)
##
u = unique(long_data$analysis)
o = as.character( u[grep("obs_", u)] )
m = as.character( u[grep("tsls_", u)] )

long_data$analysis = factor(long_data$analysis, levels = c(m,o)[length(c(m,o)):1] )

## add annotation
m = match(long_data$outcome , study_data$feature_anno$new_id)
long_data = cbind(long_data, study_data$feature_anno[m, c("metabolite","dietary_state","raw.label", "class", "subclass")])
```

## Fasting not Postprandial Forest plot

```{r, fig.height = 5, fig.width = 7}
## keep only the metabolites of interest and the wNEO observational estimates
mnames = fnotp
x = c(paste0(mnames, "_f"), paste0(mnames, "_p"))
temp_long = long_data %>% filter(analysis == "obs_wNEO" & outcome %in% x )
## redefine the analysis type
temp_long$analysis = "fasting"
w = grep("_p", temp_long$outcome)
temp_long$analysis[w] = "postprandial"
temp_long$outcome = gsub("_f","",temp_long$outcome)
temp_long$outcome = gsub("_p","",temp_long$outcome)
temp_long$outcome = factor(temp_long$outcome, levels = mnames)
##

pcol = RColorBrewer::brewer.pal(5, "Set1")[2:3]

plot_FnotP = dh_forrest_plot(data = temp_long, 
                alpha = 0.05/43, 
                analysis_type = "fasting",
                arrange_by = "pval",
                title = NA, covariates = NA, confounders = NA,
                outcome_label = "",
                exposure_label = "",
                filter_data = FALSE,
                column_4_x = "analysis",
                column_4_color = "analysis",
                col_count = 1,
                color_choices = pcol )
# plot_FnotP = plot_FnotP + theme(legend.position="bottom")
plot_FnotP

```

## Postprandial not Fasting association Forest plot

```{r, fig.height = 8, fig.width = 7}
## keep only the metabolites of interest and the wNEO observational estimates
mnames = pnotf
x = c(paste0(mnames, "_f"), paste0(mnames, "_p"))
temp_long = long_data %>% filter(analysis == "obs_wNEO" & outcome %in% x )
## redefine the analysis type
temp_long$analysis = "fasting"
w = grep("_p", temp_long$outcome)
temp_long$analysis[w] = "postprandial"
temp_long$outcome = gsub("_f","",temp_long$outcome)
temp_long$outcome = gsub("_p","",temp_long$outcome)
temp_long$outcome = factor(temp_long$outcome, levels = mnames)
##

pcol = RColorBrewer::brewer.pal(5, "Set1")[2:3]

plot_PnotF = dh_forrest_plot(data = temp_long, 
                alpha = 0.05/43, 
                analysis_type = "fasting",
                arrange_by = "pval",
                title = NA, covariates = NA, confounders = NA,
                outcome_label = "",
                exposure_label = "",
                filter_data = FALSE,
                column_4_x = "analysis",
                column_4_color = "analysis",
                col_count = 1,
                color_choices = pcol )

# plot_PnotF = plot_PnotF + theme(legend.position="bottom")
plot_PnotF
```

## Postprandial not Fasting association Forest plot

```{r, fig.height = 5, fig.width = 7}
## keep only the metabolites of interest and the wNEO observational estimates
mnames = pf_dif
x = c(paste0(mnames, "_f"), paste0(mnames, "_p"))
temp_long = long_data %>% filter(analysis == "obs_wNEO" & outcome %in% x )
## redefine the analysis type
temp_long$analysis = "fasting"
w = grep("_p", temp_long$outcome)
temp_long$analysis[w] = "postprandial"
temp_long$outcome = gsub("_f","",temp_long$outcome)
temp_long$outcome = gsub("_p","",temp_long$outcome)
temp_long$outcome = factor(temp_long$outcome, levels = mnames)
##

pcol = RColorBrewer::brewer.pal(5, "Set1")[2:3]

plot_PandF = dh_forrest_plot(data = temp_long, 
                alpha = 0.05/43, 
                analysis_type = "fasting",
                arrange_by = "pval",
                title = NA, covariates = NA, confounders = NA,
                outcome_label = "",
                exposure_label = "",
                filter_data = FALSE,
                column_4_x = "analysis",
                column_4_color = "analysis",
                col_count = 1,
                color_choices = pcol )

# plot_PandF = plot_PandF + theme(legend.position="bottom")
plot_PandF
```


```{r ,fig.height = 10, fig.width = 15}
library(patchwork)

obs_forest = ( plot_PnotF | ( plot_FnotP/plot_PandF + 
  plot_annotation(tag_levels = c("A")) +
  plot_layout(heights = c(1.5, 1) , guides = 'collect') ) ) + 
  plot_layout(guides = 'collect') + 
  plot_annotation(tag_levels = c("A")) & theme(legend.position="bottom")


obs_forest

```


```{r}
f = paste0(resultsdir, "figures/obs_fasting_postprandial_forestplot_difference_v2.pdf")
pdf(file = f, width = 10, height = 10)
obs_forest
dev.off()
```

## direction of BMI effects|associations

```{r}
w = grep("_r", obs_tsls_est$wNEO$outcome)
x = obs_tsls_est$wNEO[w,] %>% arrange(Obs_P)
x[, c("outcome", "Obs_beta","Obs_P")] %>% kable() %>% kableExtra::kable_classic()
```

## Is alanine and leucine response normally distributed ??

```{r}
x = mydata$leu_r
hist(x)
shapiro.test( na.omit(sample(x, 5000)) )
```

```{r}
plot_traits = c("ala_r","leu_r")
AA_stats = t( sapply(plot_traits, function(trait){
  out = obsmr( wdata = mydata,
            outcome = trait,
            exposure = "bmi",
            instrument = "Yengo_Weighted_GRS",
            covariates = c("visit_date","subpop","age","sex"),
            weights = "pweight",
            rnt_outcome = FALSE,
            outlier_method = "iqr",
            outlier_cutoff = 5,
            messages = TRUE)
  return(out)
}) )
##########################
## turn into a data.frame
##########################
AA_stats = as.data.frame(AA_stats)
##########################
## convert to numeric
##########################
num_cols = c(4:6, 8:9, 12:24, 27:36, 39:48, 51:74 )
for(i in num_cols){
  AA_stats[,i] = as.numeric(AA_stats[,i])
}
```


```{r}
AA_stats %>% knitr::kable() %>% kableExtra::kable_classic()
```


```{r}
obs_tsls_est$wNEO %>% filter(Obs_P < 0.05/43) %>% summarize(mean = mean(Obs_devexp_by_exposure),
                                                            min = min(Obs_devexp_by_exposure), 
                                                            max = max(Obs_devexp_by_exposure))
```

```{r}
obs_tsls_est$wNEO$outcome_type = "response"
w = grep("_f", obs_tsls_est$wNEO$outcome); obs_tsls_est$wNEO$outcome_type[w] = "fasting"
w = grep("_p", obs_tsls_est$wNEO$outcome); obs_tsls_est$wNEO$outcome_type[w] = "postprandial"

##
obs_tsls_est$wNEO %>% group_by(outcome_type) %>% filter(Obs_P < 0.05/43) %>% summarize(mean = mean(Obs_devexp_by_exposure),
                                                            min = min(Obs_devexp_by_exposure), 
                                                            max = max(Obs_devexp_by_exposure))
```

```{r}
obs_tsls_est$wNEO %>% 
  filter(outcome_type == "response") %>% 
  filter(Obs_P < 0.05/43) %>% 
  summarize(mean = mean(Obs_devexp_by_exposure) )
```


```{r}
obs_tsls_est$wNEO %>% 
  filter(outcome == "tyr_r") %>% knitr::kable() %>% kableExtra::kable_classic()
```


## Description of the Deming response data

```{r}
response_traits = traits[grep("_r", traits)]
rt_sd = apply(mydata[, response_traits], 2, function(x){ sd(x, na.rm = TRUE) }) 
rt_m = apply(mydata[, response_traits], 2, function(x){ mean(x, na.rm = TRUE) })
rt_v = apply(mydata[, response_traits], 2, function(x){ var(x, na.rm = TRUE) }) 
```


```{r}
w = which(study_data$feature_anno$ids %in% response_traits)
subc = study_data$feature_anno$class[w]
rt_v_by_class = sapply(unique(subc), function(sc){
  w = which(subc == sc)
  m = mean(rt_v[w], na.rm = TRUE)
  return(m)
})
sort(rt_v_by_class)
```

```{r}
sd = apply(mydata[, traits], 2, function(x){ sd(x, na.rm = TRUE) }) 
m = apply(mydata[, traits], 2, function(x){ mean(x, na.rm = TRUE) })
v = apply(mydata[, traits], 2, function(x){ var(x, na.rm = TRUE) }) 
##
df =  data.frame(trait = response_traits, 
                 mean = m,
                 sd = sd,
                 var = v,
                 class = study_data$feature_anno$class,
                 subclass = study_data$feature_anno$subclass, 
                 dietary_state = study_data$feature_anno$dietary_state)
```


```{r}
ndata = df %>% group_by(dietary_state, class) %>% summarize(mean_var = mean(var))
```

```{r}
ndata_wide = ndata %>% pivot_wider(names_from = dietary_state, values_from = mean_var)

t.test(ndata_wide$fasting, ndata_wide$response)

ndata_wide %>% ggplot(aes(x = fasting, y = response)) +
  geom_point() +
  geom_abline(slope = 1, intercept = 0) +
  theme_bw() # +
  # xlim(0,0.5) + ylim(0,0.5)
```




```{r}
subc = study_data$feature_anno$class
rt_v_by_class = sapply(unique(subc), function(sc){
  w = which(subc == sc)
  m = mean(rt_v[w], na.rm = TRUE)
  return(m)
})
sort(rt_v_by_class)
```



```{r, fig.width = 15, fig.height = 6}
w = which(study_data$feature_anno$ids %in% response_traits)
plot_data = data.frame(trait = response_traits, variance = rt_v, 
                       class = study_data$feature_anno$class[w],
                       subclass = study_data$feature_anno$subclass[w])

plot_data %>% ggplot(aes(x = class, y = variance)) +
  geom_boxplot(aes(fill = class)) +
  geom_jitter( width = 0.15 ) +
  geom_hline(yintercept = 0) +
  theme_bw() +
  # facet_wrap(~class, nrow = 3, scales = "free") 
  theme(axis.ticks.x = element_blank(), axis.text.x = element_blank() ) 


```

```{r}
delta = mydata$ala_p - mydata$ala_f
res = mydata$ala_r
plot(delta,res)
```
